==> 01_plant-biomass.Rmd <==
---
title: "Plant_traits"
author: "Marcela Aragon"
date: "3/1/2024"
output: 
  html_document:
    toc: true
    toc_depth: 3
    toc_float: 
     collapsed: false
     smooth_scroll: true
    df_print: paged
    code_folding: hide
    keep_md: true
editor_options: 
  chunk_output_type: console
---

This script will deal with ploting and analyzing the plant biomass in both conditioning and feedback phases;

# General settings 

## Loading packages

```{r loading packages, echo=TRUE}

#Set working directory to project directory 
setwd("./")
getwd() #ok

#data
library(knitr)
library(dplyr)
library(tibble)

#plotting
library(ggplot2)
library(ggpubr)
library(GGally)
library(RColorBrewer)
library(grDevices)
library(rempsyc) #tables

#stats
library(DHARMa)
library(lmtest)
library(fitdistrplus)
library(AICcmodavg)
library(emmeans)
library(multcompView)
library(multcomp)
library(postHoc)
library(car)
library(effsize) #Cohen's D
library(lsr)

```

## Global settings for plots 

```{r settings for plots}

# Setting up white background for plots
theme_set(theme_bw())

# Set colors
color_JA <- "#E69F00" #yellow-orange
color_SA <- "#4472C4" #blue
color_control <- "#BFBFBF" #gray
color_simulated <- "#CC3399" #pink
color_real_insects <- "#70AD47"
color_non_conditioned <-"#FFFFFF"

color_pathway_induced <- c(color_control, color_JA, color_SA) #control, JA and SA
color_experiment <- c(color_simulated, color_real_insects) #simulated and real herbivory

# Set default axis looks
axis_looks <- theme(axis.text.x = element_text(colour = "black", size = 9,
                                               face = "bold", angle=0, hjust=0.5))+ #text in X
  theme(axis.text.y = element_text(colour = "black", size = 9, face = "bold"))+ #text in y
  theme(axis.title=element_text(size=11, face = "bold"))+ #all texts
  theme(axis.title.y = element_text(margin = ggplot2::margin(t = 0, r = 10, b = 0, l = 0)))+
  theme(panel.grid.major = element_blank(), #no grid
        panel.grid.minor = element_blank()) #no grid
  #theme(legend.position="none")+ #no legend

#Get names for boxes in facet wrap
Experiment_names <- c(
                    `simulated` = "Phytohormone-treated",
                    `real_insects` = "Insect-Infested")

```


## Conditioning 

The conditioning phase for the simulated-herbivory experiment took place as part of the Family Experiment at Wageninen in between May and June from 2021. 

Briefly, in the Family experiment plants were dipped with either MeJA 0.5mM, SA 0.5mM or a Mock solution using the Berendsen et al., 2018 dipping protocol, after this plants were harvested. Shoots were dried in the oven to obtain dry shoot biomass and roots were collected for microbiome analysis. 

Some of the replicates for shoot biomass are missing as ~3-4 replicates per treatment were freeze-dried for Glucosinolate and Phytohormone-analysis. 

### Exploring data 

```{r c_phyto exploring data}

#load data
biomass <- readRDS("./plant_data/biomass.rds")

#explore 
summary(biomass)
colnames(biomass)

#quick summary
biomass %>% 
group_by(Phase, Experiment, plant_pathway_induced) %>%
  summarise(
    count = n(),
    mean = mean(shoot_biomass_mg, na.rm = TRUE),
    sd = sd(shoot_biomass_mg, na.rm = TRUE))


#subsetting df to only conditioning
conditioning <- biomass %>% 
                filter(Phase == "conditioning")


#change levels of Experiment
levels(conditioning$Experiment)
conditioning <- conditioning %>%
                mutate(Experiment = factor(Experiment, levels = c("simulated", "real_insects" )))

```

* Plants in the simulated herbivory experiment were overall bigger than those in the real insects one. 
* In both cases JA-induced plants were smaller than controls.

### Plots

Boxplot with rawdata points, colored by plant_pathway_induced and facet wraped by Experiment + colored in the upper boxes. 

```{r}

# make a boxplot
p1 <- ggplot(data = conditioning,
            mapping = aes(x = plant_pathway_induced, y = shoot_biomass_mg), colour= plant_pathway_induced)+
  geom_boxplot(mapping = aes(fill= plant_pathway_induced), position = position_dodge(width = 0.8), lwd=0.8)+
  geom_jitter(data = conditioning,
              mapping = (aes(plant_pathway_induced, shoot_biomass_mg, fill = plant_pathway_induced)),
              shape = 21, size = 1.8, color = "black",
              position = position_jitterdodge(jitter.height = .1, jitter.width = .2))+
  scale_fill_manual(values = color_pathway_induced)


p2 <- p1 + facet_wrap(~Experiment, labeller = as_labeller(Experiment_names)) +
              labs(y="Shoot biomass (mg)", x=" ") +
              ylim(400,2500)+
              axis_looks +
              theme(axis.text.x = element_blank(),
                    axis.ticks.x = element_blank())+
              theme(strip.text.x = element_text(size = 11, face = "bold"), #text in boxes
                    strip.background=element_rect(colour="black"))+ #outline of boxes
              theme(panel.spacing = unit(0.2, "lines"))+ #space in between facets
              theme(legend.position="none") #no legend, to save it better

#save plot
ggsave(
  filename = "cond_plant-biomass.svg",
  plot = last_plot(),
  path = "./Plots/01_plant-biomass",
  scale = 1.5,
  width = 1200,
  height = 700,
  units = "px",
  dpi = 300)


#clean
rm(p1,p2)
dev.off()


```

It's not straightforward to change the facet boxes by the color of the experiment. I will do this in illustrator instead. 

### Stats

First, quick check for normality to later make a one-way anova per experment using plant_pathway_induction as a factor. 

According to shapiro test data is not significantly different from a normal distribution = it's normal, and the variances between treatments are homogeneous according to the Bartlett test, for the Levene test it is significant meaning that variances are unequal. Let's check later with Dharma. 


```{r}

#check data distribution
hist(conditioning2$shoot_biomass_mg)

#two-way ANOVA
aov_biomass <- aov(shoot_biomass_mg  ~ plant_pathway_induced * Experiment,
                   data=conditioning)
summary(aov_biomass) #there is an interaction between Experiment and ppi

#checks
plot(aov_biomass, which = 2) #q-q plot, there are some outliers
qqPlot(aov_biomass$residuals, #also in this q-q plot, let's split it by experiment instead 
  id = T)
plot(aov_biomass, which = 3)

```


Simulated Herbivory

```{r}
x <- simu %>% group_by(plant_pathway_induced) %>%
  summarise(
    n = n(),
    mean = mean(shoot_biomass_mg, na.rm = TRUE))


summarise(
    n = n(),
    mean = mean(concentration, na.rm = TRUE))
```

```{r}
#quick summary
phyto %>% group_by(Experiment, soil_inocula) %>% summarise(n=n()) #not balanced

#with interaction
lm <- glm(shoot_biomass_mg ~ soil_inocula*herbivory,
          family=gaussian,
          data=phyto)

lm.anova <- car::Anova(lm, type="III", test.statistic="LR")
simulateResiduals(lm, plot=T) #ok


#pairiwse differences
emmeans_herb <- emmeans(lm, pairwise~herbivory) #differences between NI and caterpillar

#Save tables 
. <- c("Soil inocula (S)", "Herbivory (H)", "S*H")
lm_table <- cbind (., as.data.frame(lm.anova))
lm_table$`Pr(>Chisq)` <- round(lm_table$`Pr(>Chisq)`, 3)
table <- nice_table(lm_table,
              title = c("Table X. GLM for plant biomass in Feedback phase - Phyto"),
              note = "shoot_biomass_mg  ~ soil_inocula*herbivory, GLM-Gaussian")

table <- table %>%
         flextable::colformat_num(big.mark = "", digits = 3) 

flextable::save_as_docx(table, path = "./Tables/01_plant-biomass/feedback_lm_biomass_phyto.docx")
```


```{r}
#one-way ANOVA: Simulated
simu <- subset(conditioning, Experiment == "simulated") %>% droplevels()

aov_simulated <- glm(shoot_biomass_mg  ~ plant_pathway_induced,
                     family=gaussian,
                   data=simu)

lm.anova <- car::Anova(aov_simulated, type="III", test.statistic="LR")

simulateResiduals(aov_simulated, plot=T) #ok

#Save tables 
. <- c("Treatment")
lm_table <- cbind (., as.data.frame(lm.anova))
lm_table$`Pr(>Chisq)` <- round(lm_table$`Pr(>Chisq)`, 3)
table <- nice_table(lm_table,
              title = c("Table X. GLM for plant biomass in Conditioning phase - Phyto"),
              note = "shoot_biomass_mg  ~ treatment, GLM-Gaussian")

table <- table %>%
         flextable::colformat_num(big.mark = "", digits = 3) 

flextable::save_as_docx(table, path = "./Tables/01_plant-biomass/conditioning_lm_biomass_phyto.docx")

###

#post-hoc test
emmeans <- emmeans(aov_simulated, "plant_pathway_induced", data=simu)
posthoc <- cld(emmeans, letters="letters", adjust="tukey")
plot(emmeans, comparisons = TRUE) #check 

#make tables
#Anova
. <- c("Plant pathway induced", "Residuals")
aov_table <- cbind (.,as.data.frame(Anova(aov_simulated)))
table <- nice_table(aov_table,
              title = c("Table X. One-way ANOVA for Plant Biomass", "Conditioning phase - Simulated Herbivory"),
              note = "Signif. codes:  0 ‘***’ 0.001 ‘**’ 0.01 ‘*’ 0.05 ‘.’ 0.1 ‘ ’ 1")

flextable::save_as_docx(table, path = "./Tables/01_plant-biomass/cond_aov_plant-biomass_simulated.docx")

#Post-Hoc
ph_table <- nice_table(as.data.frame(posthoc),
              title = c("Table X. Post-hoc for Plant Biomass", "Conditioning phase - Simulated Herbivory"),
              note = "Post-hoc: Tukey")

flextable::save_as_docx(ph_table, path = "./Tables/01_plant-biomass/cond_ph_plant-biomass_simulated.docx")

#clean
rm(emmeans, posthoc, ph_table, ., aov_table, table)

```

Real Herbivory

```{r}
x <- real %>% group_by(plant_pathway_induced) %>%
  summarise(
    n = n(),
    mean = mean(shoot_biomass_mg, na.rm = TRUE))
```

```{r}
#one-way ANOVA: Real Insects
real <- subset(conditioning, Experiment == "real_insects") %>% droplevels()

#quick summary
real %>% group_by(Experiment, plant_pathway_induced) %>% summarise(n=n())

aov_real <- glm(shoot_biomass_mg  ~ plant_pathway_induced,
                family=gaussian,
                   data=real)

lm.anova <- car::Anova(aov_real, type="II", test.statistic="LR")

simulateResiduals(aov_real, plot=T) #ok

#Save tables 
. <- c("Treatment")
lm_table <- cbind (., as.data.frame(lm.anova))
lm_table$`Pr(>Chisq)` <- round(lm_table$`Pr(>Chisq)`, 3)
table <- nice_table(lm_table,
              title = c("Table X. GLM for plant biomass in Conditioning phase - Insects"),
              note = "shoot_biomass_mg  ~ treatment, GLM-Gaussian")

table <- table %>%
         flextable::colformat_num(big.mark = "", digits = 3) 

flextable::save_as_docx(table, path = "./Tables/01_plant-biomass/conditioning_lm_biomass_insects.docx")

#post-hoc test
emmeans <- emmeans(aov_real, "plant_pathway_induced", data=real)
posthoc <- cld(emmeans, letters="letters", adjust="tukey")
plot(emmeans, comparisons = TRUE) #check 

#make tables
#Anova
. <- c("Plant pathway induced", "Residuals")
aov_table <- cbind (.,as.data.frame(Anova(aov_real)))
table <- nice_table(aov_table,
              title = c("Table X. One-way ANOVA for Plant Biomass", "Conditioning phase - Insect Herbivory"),
              note = "Signif. codes:  0 ‘***’ 0.001 ‘**’ 0.01 ‘*’ 0.05 ‘.’ 0.1 ‘ ’ 1")

flextable::save_as_docx(table, path = "./Tables/01_plant-biomass/cond_aov_plant-biomass_real.docx")

#Post-Hoc
ph_table <- nice_table(as.data.frame(posthoc),
              title = c("Table X. Post-hoc for Plant Biomass", "Conditioning phase - Insect Herbivory"),
              note = "Post-hoc: Tukey")

flextable::save_as_docx(ph_table, path = "./Tables/01_plant-biomass/cond_ph_plant-biomass_real.docx")

#clean
rm(ph_table, aov_table, table, emmeans, posthoc, real, simu, aov_real, aov_simulated)

```


## Feedback 

**[Explanation about experimental design]**

In the feedback phase I'm interested in knowing the effect of both the type of soil inocula and the type of herbivory on the plant's biomass. Thus,the data will be analyzed with a two-way ANOVA using 'soil_inocula' and 'herbivory' as factors. One analysis will be made per experiment.

### Exploring data 

First, let's have a quick look to the data 
```{r}

#subsetting df to only feedback
feedback <- biomass %>% 
                filter(Phase == "feedback")

#change levels of Experiment & Herbivory
levels(feedback$Experiment)
levels(feedback$herbivory)
levels(feedback$soil_inocula)

feedback <- feedback %>%
                mutate(Experiment = factor(Experiment, levels = c("simulated", "real_insects" )))%>%
                mutate(herbivory = factor(herbivory, levels = c("Without", "Chewer", "Aphid")))

feedback <- feedback %>%
                mutate(herbivory = factor(herbivory, labels = c("uninfested", "Mamestra", "Myzus")))%>%
                filter(!is.na(shoot_biomass_mg))%>% #remove NAs
                droplevels()

str(feedback)

#finally, delete not informative columns
feedback <- feedback %>% dplyr::select(-Phase, -plant_pathway_induced, -Stress, -Leaf_area)

feedback_full <- feedback
str(feedback_full)

feedback <- feedback %>%  subset(soil_inocula != "non-conditioned") %>% droplevels()
str(feedback)


```


### Plots

Boxplot with rawdata points, colored by plant_pathway_induced and facet wraped by Experiment + colored in the upper boxes. 

```{r}

# make a boxplot
biomass_p <- ggplot(data = subset(feedback, Experiment == "simulated"),
       mapping = aes(x = soil_inocula, y = shoot_biomass_mg), colour= soil_inocula)+
  geom_boxplot(mapping = aes(fill= soil_inocula), position = position_dodge(width = 0.8), lwd=0.8)+
  geom_jitter(data = subset(feedback, Experiment == "simulated"),
              mapping = (aes(soil_inocula, shoot_biomass_mg, fill = soil_inocula)),
              shape = 21, size = 1.8, alpha=0.9, color = "black",
              position = position_jitterdodge(jitter.height = .1, jitter.width = .2))+
  scale_fill_manual(values = color_pathway_induced)+
  axis_looks+
  ylim(100,2000)+
  theme(legend.position="none")+
  facet_wrap(~herbivory)+
  theme(panel.spacing = unit(0, "lines"))


biomass_i <- ggplot(data = subset(feedback, Experiment == "real_insects"),
       mapping = aes(x = soil_inocula, y = shoot_biomass_mg), colour= soil_inocula)+
  geom_boxplot(mapping = aes(fill= soil_inocula), position = position_dodge(width = 0.8), lwd=0.8)+
  geom_jitter(data = subset(feedback, Experiment == "real_insects"),
              mapping = (aes(soil_inocula, shoot_biomass_mg, fill = soil_inocula)),
              shape = 21, size = 1.8, alpha=0.9,color = "black",
              position = position_jitterdodge(jitter.height = .1, jitter.width = .2))+
  scale_fill_manual(values = color_pathway_induced)+
  axis_looks+
  ylim(100,2000)+
  theme(legend.position="none")+
  facet_wrap(~herbivory)+
  theme(panel.spacing = unit(0, "lines"))


# Combine the two plots
library(patchwork)
combined_plot <- biomass_p + biomass_i

#save plot
mapply(function(x)
  ggsave(
  plot = last_plot(),
  filename =x,
  path = "./Plots/01_plant-biomass",
  scale = 2,
  width = 1800,
  height = 500,
  units = "px",
  dpi = 300),
x=c("feedback_biomass.svg",
    "feedback_biomass.png"))

```

Old plot
```{r}
p1 <- ggplot(data = feedback,
            mapping = aes(x = soil_inocula, y = shoot_biomass_mg), colour= soil_inocula)+
  geom_boxplot(mapping = aes(fill=soil_inocula), position = position_dodge(width = 0.8), lwd=0.8)+
  geom_jitter(data = feedback,
              mapping = (aes(x = soil_inocula, y = shoot_biomass_mg, fill = soil_inocula)),
              shape = 21, size = 1.4, color = "black",
              position = position_jitterdodge(jitter.height = .1, jitter.width = .2))+
  scale_fill_manual(values = color_pathway_induced)+
  axis_looks


#Get names for boxes in facet wrap
Experiment_names <- c(
                    `simulated` = "Simulated Herbivory",
                    `real_insects` = "Insect Herbivory",
                    `uninfested` = "Not infested",
                    `Mamestra` = "M. brassicae caterpillars",
                    `Myzus` = "M. persicae aphids")


p2 <- p1 + facet_wrap(herbivory~Experiment, ncol=2,
                      labeller = as_labeller(Experiment_names)) +
              labs(y="Shoot biomass (mg)", x=" ") +
              #ylim(400,2500)+
              axis_looks +
              theme(axis.text.x = element_blank(),
                    axis.ticks.x = element_blank())+
              theme(strip.text.x = element_text(size = 7, face = "bold"), #text in boxes
                    strip.background=element_blank())+ #outline of boxes
              theme(panel.spacing = unit(0.0, "lines"))+ #space in between facets
              theme(legend.position="none") #no legend, to save it better


#save plot
ggsave(
  filename = "feedback_plant-biomass.svg",
  plot = last_plot(),
  path = "./Plots/01_plant-biomass",
  scale = 1.5,
  width = 600,
  height = 1600,
  units = "px",
  dpi = 300)


#clean
rm(p1,p2)
dev.off()

```


### Stats

```{r}

#---- Phyto ----
phyto <- subset(feedback, Experiment == "simulated")

#quick summary
phyto %>% group_by(Experiment, soil_inocula, herbivory) %>% summarise(n=n()) #not balanced

#with interaction
lm <- glm(shoot_biomass_mg ~ soil_inocula*herbivory,
          family=gaussian,
          data=phyto)

lm.anova <- car::Anova(lm, type="III", test.statistic="LR")
simulateResiduals(lm, plot=T) #ok

#pairiwse differences
#emmeans_herb <- emmeans(lm, pairwise~herbivory) #differences between NI and caterpillar

#Save tables 
. <- c("Soil inocula (S)", "Herbivory (H)", "S*H")
lm_table <- cbind (., as.data.frame(lm.anova))
lm_table$`Pr(>Chisq)` <- round(lm_table$`Pr(>Chisq)`, 3)
table <- nice_table(lm_table,
              title = c("Table X. GLM for plant biomass in Feedback phase - Phyto"),
              note = "shoot_biomass_mg  ~ soil_inocula*herbivory, GLM-Gaussian")

table <- table %>%
         flextable::colformat_num(big.mark = "", digits = 3) 

flextable::save_as_docx(table, path = "./Tables/01_plant-biomass/feedback_lm_biomass_phyto.docx")

#---- Insects ----
insects <- subset(feedback, Experiment == "real_insects")
insects %>% group_by(Experiment, soil_inocula, herbivory) %>% summarise(n=n()) #balanced

lm.i <- glm(shoot_biomass_mg ~ soil_inocula*herbivory,
            family=gaussian,
            data=insects)

lm.i.anova <- car::Anova(lm.i, type="II", test.statistic="LR")
simulateResiduals(lm.i, plot=T) #ok

#pairiwse differences
emmeans_herb <- emmeans(lm.i, pairwise~herbivory) #differences between NI and caterpillar

#insects
lm.i_table <- cbind (., as.data.frame(Anova(lm.i)))
lm.i_table$`Pr(>Chisq)` <- round(lm.i_table$`Pr(>Chisq)`, 3)
table <- nice_table(lm.i_table,
              title = c("Table X. GLM for plant biomass in Feedback phase - Insects"),
              note = "shoot_biomass_mg  ~ soil_inocula*herbivory, GLM-Gaussian")

table <- table %>%
         flextable::colformat_num(big.mark = "", digits = 3) 

flextable::save_as_docx(table, path = "./Tables/01_plant-biomass/feedback_lm_biomass_insects.docx")


```


Old stats
```{r}

#check data distribution
hist(feedback$shoot_biomass_mg) #normal

#lm with 3 factors
lm <- lm(shoot_biomass_mg  ~ soil_inocula*herbivory*Experiment,
                   data=feedback)

summary(lm) 
Anova(lm)

#checks
plot(lm, which = 2) #q-q plot ok
simulateResiduals(lm, plot=T) #ok

#save table
. <- c("Soil inocula (S)", "Herbivory (H)", "Experiment (E)", "S*E", "S*E", "H*E", "S*H*E", "Residuals")
lm_table <- cbind (., as.data.frame(Anova(lm)))
table <- nice_table(as.data.frame(Anova(lm)),
              title = c("Table X. Linear Model for plant biomass in Feedback phase"),
              note = "shoot_biomass_mg  ~ soil_inocula*herbivory*Experiment")

flextable::save_as_docx(table, path = "./Tables/01_plant-biomass/feedback_lm_biomass.docx")


#post-hoc test
emmeans <- emmeans(lm, "herbivory", data=feedback)
posthoc <- pairs(emmeans, adjust="tukey")
plot(emmeans, comparisons = TRUE) #check 

#Post-Hoc
ph_table <- nice_table(as.data.frame(posthoc),
              title = "Table X. Post-hoc for Plant Biomass in Feedback phase",
              note = "Post-hoc: Tukey")

flextable::save_as_docx(ph_table, path = "./Tables/01_plant-biomass/feedback_ph_plant-biomass.docx")


```


### Cohen's D

```{r}

library(effsize)
library(rstatix)

feedback_full
phyto <- subset(feedback_full, Experiment == "simulated")
insect <- subset(feedback_full, Experiment == "real_insects")
levels(phyto$soil_inocula)

#Phyto
ctr <- phyto %>% filter(soil_inocula %in% c("non-conditioned", "control-conditioned"))

control <- cohen.d(ctr$shoot_biomass_mg[ctr$soil_inocula == "control-conditioned"],
                     ctr$shoot_biomass_mg[ctr$soil_inocula == "non-conditioned"],
                     paired = FALSE, hedges.correction = TRUE)

ctr.i <- insect %>% filter(soil_inocula %in% c("non-conditioned", "control-conditioned"))

control <- cohen.d(ctr.i$shoot_biomass_mg[ctr.i$soil_inocula == "control-conditioned"],
                     ctr.i$shoot_biomass_mg[ctr.i$soil_inocula == "non-conditioned"],
                     paired = FALSE, hedges.correction = TRUE)


```
