==> 03_glucosinolates.Rmd <==
---
title: "03_glucosinolates.Rmd"
author: "Marcela Aragon"
date: "15/1/2024"
output: 
  html_document:
    toc: true
    toc_depth: 3
    toc_float: 
     collapsed: false
     smooth_scroll: true
    df_print: paged
    code_folding: hide
    keep_md: true
editor_options: 
  chunk_output_type: console
---

This script will deal with ploting and analyzing the concentration of X different glucosinolates in both conditioning and feedback phases;

# General settings 

## Loading packages

```{r loading packages, echo=TRUE}

#Set working directory to project directory 
setwd("./")
getwd() #ok

#data
library(knitr)
library(dplyr)
library(tibble)

#plotting
library(ggplot2)
library(ggpubr)
library(GGally)
library(RColorBrewer)
library(grDevices)
library(rempsyc) #tables

#stats
library(DHARMa)
library(lmtest)
library(fitdistrplus)
library(AICcmodavg)
library(emmeans)
library(multcompView)
library(multcomp)
library(postHoc)
library(car)
library(effsize) #Cohen's D
library(lsr)

library(ropls) #PSLDA
library(mixOmics) #PSLDA
library(PERMANOVA)
library(plsdepot)
library(santaR)
library(vegan)

```

## Global settings for plots 

```{r settings for plots}

# Setting up white background for plots
theme_set(theme_bw())

# Set colors
color_JA <- "#E69F00" #yellow-orange
color_SA <- "#4472C4" #blue
color_control <- "#BFBFBF" #gray
color_simulated <- "#CC3399" #pink
color_real_insects <- "#70AD47"
color_non_conditioned <-"#FFFFFF"

color_pathway_induced <- c(color_control, color_JA, color_SA) #control, JA and SA
color_experiment <- c(color_simulated, color_real_insects) #simulated and real herbivory

# Set default axis looks
axis_looks <- theme(axis.text.x = element_text(colour = "black", size = 9,
                                               face = "bold", angle=0, hjust=0.5))+ #text in X
  theme(axis.text.y = element_text(colour = "black", size = 9, face = "bold"))+ #text in y
  theme(axis.title=element_text(size=11, face = "bold"))+ #all texts
  theme(axis.title.y = element_text(margin = ggplot2::margin(t = 0, r = 10, b = 0, l = 0)))+
  theme(panel.grid.major = element_blank(), #no grid
        panel.grid.minor = element_blank()) #no grid
  #theme(legend.position="none")+ #no legend

#Get names for boxes in facet wrap
Experiment_names <- c(
                    `simulated` = "Simulated Herbivory",
                    `real_insects` = "Insect Herbivory")

```

## Conditioning 

The conditioning phase for the simulated-herbivory experiment took place as part of the Family Experiment at Wageninen in between May and June from 2021. And the insect-herbivory experiment took place between *X and X from 2021.*

From the 16-17 total replicates in the conditioning phase, 4 plants per treatment were randomly allocated to be used for metabolite analysis. For this, shoot biomass was cut from the root and flash-froze in liquid nitrogen. Samples were freeze-dried for 72 hours (*equipment*) and grinded to fine powder at the UNIFARM facilities in Wageningen (big extruder in grinding room). Per sample, 15-30 mg (*I need to check this*) were weighed individually and placed in a special tray (*get the name*) to be quantified by LC-MS/MS in Jenna by Michael Reichelt.

Text by Michael Reichelt: 

"Glucosinolate analysis by HPLC-UV

Glucosinolates were analysed by HPLC-UV as described in Burow et.al. (2006). 6-25 mg of lyophilized root material were extracted with 1mL methanol solution (v:V) containing 50nmol of 4-hydroxybenzyl glucosinolate. A 500uL aliquot of the raw extract was loaded onto DEAE Sephadex A 25 columns and treated with arylsulfatase for desulfation (Sigma-Aldrich) (Brown et al. 2003). The eluted desulfoglucosinolates were separated using high performance liquid chromatography (Agilent 1100 HPLC system, Agilent Technologies) on a reversed phase C-18 column (Nucleodur Sphinx RP, 250 x 4.6 mm, 5µm, Machrey-Nagel, Düren, Germany) with a water (A)-acetonitrile (B) gradient (0-1 min, 1.5% B; 1-6 min, 1.5-5% B; 6-8 min, 5-7% B; 8-18 min, 7-21% B; 18-23 min, 21-29% B; 23-23.1 min, 29-100% B; 23.1-24min 100% B and 24.1-28 min 1.5% B; flow 1.0 mL min-1). Detection was performed with a photodiode array detector and peaks were integrated at 229 nm. Desulfated  glucosinolates were  identified  by  comparison  of  retention  time  and  UV spectra  to  those  of  purified  standards  previously  extracted from Arabidopsis thaliana (Brown et al. 2003) or by analysis of the desulfoglucosinolate extracts on an LC-ESI-Ion-Trap-mass spectrometer (Esquire6000, Bruker Daltonics). We used the following molar response factors for quantification of individual Gls relative to  the  internal  standard,  4-hydroxybenzyl glucosinolate: aliphatic Gls 2.0, indole Gls 0.5 (Burow et al. 2006). The following glucosinolates were detected in the samples: 3-methylsulfinylpropyl glucosinolate (3MSOP), 4-methylsulfinylbutyl glucosinolate (4MSOB), 2-propenyl glucosinolate (allyl), 4-hydroxy-indol-3-ylmethyl glucosinolate (4OHI3M), indol-3-ylmethyl glucosinolate (I3M), 4-methoxy-indol-3-ylmethyl glucosinolate (4MOI3M), and 1-methoxy-indol-3-ylmethyl glucosinolate (1MOI3M)."



### Exploring data 


```{r}

#load data
gluco <- readRDS("./plant_data/glucosinolates.rds")

#change levels of Experiment
levels(gluco$Experiment)
gluco <- gluco %>%
                mutate(Experiment = factor(Experiment,
                                           levels = c("simulated", "real_insects" )))%>%
                mutate(plant_pathway_induced = factor(plant_pathway_induced,
                                                      levels=c("Control", "JA", "SA")))
#remove sample
gluco <-gluco %>%
        subset(ID != "159")%>% #dry sample, no info
        droplevels()

#### classify by Indolic or Aliphatic ####
indolic <- c("Glu_4OHI3M", "Glu_I3M", "Glu_1MOI3M", "Glu_4MOI3M") 
aliphatic <- c("Glu_Allyl", "Glu_3MSOP","Glu_4MSOB")
#make a small df
values <- c(indolic, aliphatic)
#Create a vector of names, repeating the vector names accordingly
names <- c(rep("indolic", length(indolic)), rep("aliphatic", length(aliphatic)))
# Create the data frame
key_glucosinolates <- data.frame(type = names, glucosinolate = values)
#clean
rm(values,names)
#### ####

gluco <- gluco %>%
          relocate(Glu_Allyl,Glu_3MSOP,Glu_4MSOB, .after=weight_mg) %>% 
          relocate(Glu_4OHI3M,Glu_I3M,Glu_1MOI3M, Glu_4MOI3M, .after=Glu_4MSOB)%>% 
          mutate(Glu_alyphatic=rowSums(.[11:13]))%>%
          mutate(Glu_indolic=rowSums(.[14:17]))


#make long format for plotting
long <- gluco %>% 
              tidyr::pivot_longer(cols = c(11:20), 
              names_to = "glucosinolate",
              values_to = "concentration",
              values_drop_na = TRUE)

#making new phytohormone column as factor
long$glucosinolate <- as.factor(long$glucosinolate)

# Reordering by indolic and aliphatic
#ordered_glucosinolates <- c("Glu_total", "Glu_I3M", "Glu_4OHI3M", "Glu_1MOI3M", "Glu_4MOI3M", "Glu_3MSOP", "Glu_4MSOB", "Glu_Allyl")

#long$glucosinolate <-factor(long$glucosinolate, 
#                             levels = ordered_glucosinolates)

#subsetting df to only conditioning
conditioning_long <- long %>% 
                filter(Phase == "conditioning")%>%
                filter(glucosinolate != "Glu_4OHI3M") #mostly 0's

conditioning_gluco <- gluco %>%
                      dplyr::filter(Phase == "conditioning")%>%
                      dplyr::select(-ends_with('4OHI3M')) #mostly 0's

#first filter by those glucosinolates
long_cond <- conditioning_long %>% filter(
  glucosinolate %in% c("Glu_alyphatic","Glu_indolic")&
    Phase == "conditioning")

#Now make a summary df
df_GLS_p <- long_cond %>% 
            group_by(Experiment, plant_pathway_induced, glucosinolate) %>%
  summarise(
    n = n(),
    mean = mean(concentration, na.rm = TRUE),
    sd = sd(concentration, na.rm = TRUE)) %>%
    mutate(se = sd/sqrt(n))

```

*Now I have 3 df, gluco, conditioning_long and conditioning_gluco 


###Plot indolic, aliphatic and total
```{r}



#Finally, plot it
ggplot(data = df_GLS_p,
       mapping = aes(x = glucosinolate, y = mean), colour= plant_pathway_induced)+
  geom_bar(aes(fill = plant_pathway_induced), stat = "identity",
           position = position_dodge(0.8), width = 0.7, color = "black")+
  geom_errorbar(aes(ymin = mean-se, ymax = mean+se, group = plant_pathway_induced),
                width = 0.2, position = position_dodge(0.8))+
  geom_jitter(data = long_cond,
              mapping = (aes(glucosinolate, concentration,
                             fill = plant_pathway_induced)),
              shape = 21, size = 1.5, color = "black",
              position = position_jitterdodge(jitter.height = .1, jitter.width = .2))+
  scale_fill_manual(values = color_pathway_induced)+
  axis_looks+
  facet_wrap(~Experiment)

#save plot
mapply(function(x)
  ggsave(
  plot = last_plot(),
  filename =x,
  path = "./Plots/03_glucosinolates",
  scale = 1.8,
  width = 1300,
  height = 600,
  units = "px",
  dpi = 300),
x=c("cond_barplot_aliphatic_indolic.svg",
    "cond_barplot_aliphatic_indolic.png"))

```

* Values look similar between experiments in the conditioning phase
* There are slightly less replicates for SA

###Stats per GLS


```{r}

#Phyto
phyto.i <- subset(long_cond, Experiment == "simulated" & glucosinolate == "Glu_alyphatic")

phyto.i %>% group_by(Experiment, plant_pathway_induced) %>% summarise(n=n()) #balanced N=4

lm <- glm(log(concentration) ~ plant_pathway_induced,
          family=gaussian,
          data=phyto.i)

lm_table <- car::Anova(lm, type="II", test.statistic="LR")
simulateResiduals(lm, plot=T) #ok

#save result
. <- c("Treatment(T)", "Glucosinolate(G)", "T*G")
lm_table <- cbind (., as.data.frame(lm_table))
lm_table$`Pr(>Chisq)` <- round(lm_table$`Pr(>Chisq)`, 3)
table <- nice_table(lm_table,
              title = c("Table X. GLM for GLS in conditioning phase - Phyto"),
              note = "~ Treatment*glucosinolate type, GLM-Gaussian")

table <- table %>%
         flextable::colformat_num(big.mark = "", digits = 3) 

flextable::save_as_docx(table, path = "./Tables/03_glucosinolates/conditioning_phyto.docx")

#emeans
emmeans <- emmeans(lm, ~plant_pathway_induced)
pair <- contrast(emmeans, method="pairwise", adjust="bonferroni")
cld <- cld(emmeans, Letters="letters", adjust="tukey")




#Insects
#Phyto
insects <- subset(long_cond, Experiment == "real_insects" & glucosinolate == "Glu_indolic")
insects %>% group_by(plant_pathway_induced) %>% summarise(n=n()) #balanced N=4


lm <- glm(log(concentration) ~ plant_pathway_induced,
          family=gaussian,
         data=insects)

#GLS
lm_table <- car::Anova(lm, type="III", test.statistic="LR")
simulateResiduals(lm, plot=T)

#save result
. <- c("Treatment(T)", "Glucosinolate(G)", "T*G")
lm_table <- cbind (., as.data.frame(lm_table))
lm_table$`Pr(>Chisq)` <- round(lm_table$`Pr(>Chisq)`, 3)
table <- nice_table(lm_table,
              title = c("Table X. GLM for GLS in conditioning phase - Insects"),
              note = "~ Treatment*glucosinolate type, GLM-Gaussian")

table <- table %>%
         flextable::colformat_num(big.mark = "", digits = 3) 

flextable::save_as_docx(table, path = "./Tables/03_glucosinolates/conditioning_insects.docx")


#emmeans
emmeans <- emmeans(lm, ~plant_pathway_induced)
pair <- contrast(emmeans, method="pairwise", adjust="bonferroni")
pair
cld <- cld(emmeans, Letters="letters", adjust="tukey")
cld

```

###Stats

```{r}

#Phyto
phyto <- subset(long_cond, Experiment == "simulated")

phyto %>% group_by(Experiment, plant_pathway_induced) %>% summarise(n=n()) #balanced N=8

lm <- glm(log(concentration) ~ plant_pathway_induced*glucosinolate,
          family=gaussian,
         data=phyto)

lm_table <- car::Anova(lm, type="II", test.statistic="LR")
simulateResiduals(lm, plot=T) #ok

#save result
. <- c("Treatment(T)", "Glucosinolate(G)", "T*G")
lm_table <- cbind (., as.data.frame(lm_table))
lm_table$`Pr(>Chisq)` <- round(lm_table$`Pr(>Chisq)`, 3)
table <- nice_table(lm_table,
              title = c("Table X. GLM for GLS in conditioning phase - Phyto"),
              note = "~ Treatment*glucosinolate type, GLM-Gaussian")

table <- table %>%
         flextable::colformat_num(big.mark = "", digits = 3) 

flextable::save_as_docx(table, path = "./Tables/03_glucosinolates/conditioning_phyto.docx")

#emeans
emmeans <- emmeans(lm, ~plant_pathway_induced*glucosinolate)
pair <- contrast(emmeans, method="pairwise", adjust="bonferroni")
cld <- cld(emmeans, Letters="letters", adjust="tukey")




#Insects
#Phyto
insects <- subset(long_cond, Experiment == "real_insects")

lm <- glm(log(concentration) ~ plant_pathway_induced*glucosinolate,
          family=gaussian,
         data=insects)

#GLS
lm_table <- car::Anova(lm, type="II", test.statistic="LR")
simulateResiduals(lm, plot=T)

#save result
. <- c("Treatment(T)", "Glucosinolate(G)", "T*G")
lm_table <- cbind (., as.data.frame(lm_table))
lm_table$`Pr(>Chisq)` <- round(lm_table$`Pr(>Chisq)`, 3)
table <- nice_table(lm_table,
              title = c("Table X. GLM for GLS in conditioning phase - Insects"),
              note = "~ Treatment*glucosinolate type, GLM-Gaussian")

table <- table %>%
         flextable::colformat_num(big.mark = "", digits = 3) 

flextable::save_as_docx(table, path = "./Tables/03_glucosinolates/conditioning_insects.docx")


#emmeans
emmeans <- emmeans(lm, ~plant_pathway_induced*glucosinolate)
pair <- contrast(emmeans, method="pairwise", adjust="bonferroni")
pair
cld <- cld(emmeans, Letters="letters", adjust="tukey")
cld

```






### PLSDA

Stands for Partial Least Squares Discriminant Analysis; a type of multivariate dimensionality reduction analysis. It can be thought as a "supervised" version of a PCA. The general approach is to analyze all phytohormones in a single analysis to check for patterns.  

https://omicsforum.ca/t/what-is-the-meaning-of-components-in-pls-da/732/3

```{r}

#The following code was written by Els van de Zande, thank you Els!

View(conditioning_gluco)

conditioning <- conditioning_gluco

str(conditioning)
matrixdata<-conditioning[,11:17]# keep only values
matrixdata <- as.matrix(matrixdata[,])# make it numeric
metadata<-conditioning[,c("Experiment", "plant_pathway_induced")]
metadata_number <- metadata %>% #change from characters to numbers
                     mutate(Experiment = case_when(Experiment == 'simulated' ~ '1',
                                            TRUE ~ '2')) %>%
                     mutate(plant_pathway_induced = case_when(plant_pathway_induced == 'Control' ~ '1',
                                                       plant_pathway_induced == 'JA' ~ '2',
                                                       TRUE ~ '3'))
MetaData <-data.matrix(metadata_number) #makes a numeric matrix

#PLS-DA
plsda <-opls(matrixdata, MetaData, predI=2)
plsda_center <-opls(matrixdata,MetaData, scaleC = "center", predI=2)
plsda_pareto <-opls(matrixdata,MetaData, scaleC = "pareto", predI=2)#removes the relative importance of large values
                                                          #but keeps the data structure intact.
#check
plot(plsda_pareto)
plsda_pareto@scoreMN
plot(plsda_pareto,
     typeVc = "x-score",
     parAsColFcVn = MetaData[,"plant_pathway_induced"],
     parLabVc = as.character(MetaData[,"plant_pathway_induced"]))

#split by factors
plant_pathway_Factor <- MetaData[,"plant_pathway_induced"]
experiment_Factor <- MetaData[,"Experiment"]

plsda_pareto2 <-opls(matrixdata,plant_pathway_Factor, scaleC = "pareto", permI=1000, predI=2) #plant pathway induced
plsda_pareto3 <-opls(matrixdata,experiment_Factor, scaleC = "pareto", permI=1000, predI=2) #Experiment 

#get scores to plot them
scores<-plsda_pareto@scoreMN
scores_df<-as.data.frame(scores) #make new df 

scores_df$p1 <- as.numeric(scores_df$p1)
scores_df$p2 <- as.numeric(scores_df$p2)

scores_df2 <- scores_df %>% #add info
              mutate(Experiment = as.numeric(metadata$Experiment))%>%
              mutate(plant_pathway_induced = as.numeric(metadata$plant_pathway_induced))
              
#Plot the scores using ggplot2
scoreplot<-ggplot(scores_df2, aes(x = p1, y = p2, fill = factor(plant_pathway_induced),
                                  shape = factor(Experiment))) +
  geom_point(aes(colour=factor(plant_pathway_induced), fill=factor(plant_pathway_induced)),
              size = 4, alpha=0.8)+
  labs(x = "t1(90%)", y = "t2(5%)",
       color="Plant pathway induced", shape="Experiment", fill=element_blank()) +
  scale_fill_manual(values =color_pathway_induced) +  #Customize color scheme for Treatment
  scale_color_manual(values =color_pathway_induced)
  scale_shape_manual(values = c(16,17)) +  #Customize shape for Herbivore
  axis_looks

#now add loadings
loadings<-plsda_pareto@loadingMN
loadings_df<-as.data.frame(loadings)

loadings_df$p1 <- as.numeric(loadings_df$p1)
loadings_df$p2 <- as.numeric(loadings_df$p2)

#makes plot with arrows marking direction
rownames(loadings)
glucosinolate_names <- c("Total","3MSOP","4MSOB","Sinigrin","I3M","4MOI3M","1MOI3M")

Biplot <-scoreplot+
         geom_hline(yintercept=0, linetype='dotted', col = 'gray20',linewidth = 0.2 )+
         geom_vline(xintercept=0, linetype='dotted', col = 'gray20', linewidth = 0.2)+
         geom_segment(data = loadings_df, aes(x = 0, y = 0, xend = (p1*8), yend = (p2*8)),
                      arrow = arrow(length = unit(0.5, "picas"), type="closed"),
                      colour = "black", linewidth = 0.5, inherit.aes = FALSE)+
         geom_text(aes(x = (p1*8), y = (p2*8), label = glucosinolate_names,
                angle = 0.45, hjust = -.1, vjust = .2),
                 data = loadings_df,
                 colour = "black", size=4, inherit.aes = FALSE)+
        #ylim(-60,60)+
        #xlim(-100,100)+
        axis_looks
         #theme(legend.position = "none")

#save plot
ggsave(
  filename = "cond_glucosinolates_PLSDA.svg",
  plot = Biplot,
  path = "./Plots/03_glucosinolates",
  scale = 1.5,
  width = 1200,
  height = 700,
  units = "px",
  dpi = 300)


#Let's also get the VIP scores 

#extract VIP (Variable's importance) scores and CV-value
VIPmatrixdataPLSDA<-getVipVn(plsda_pareto, orthoL = FALSE)

VIP <- as.data.frame(VIPmatrixdataPLSDA)
VIP <- VIP %>%
        arrange(VIP)%>%
        rename(VIP = VIPmatrixdataPLSDA)%>%
        mutate(glucosinolate=rownames(.))

VIP$glucosinolate <- as.factor(VIP$glucosinolate)

#plot it to have it next to the plot
p_VIP <- VIP %>%
         mutate(glucosinolate=factor(glucosinolate, levels=glucosinolate)) %>%
         ggplot(aes(x = glucosinolate, y = VIP)) +
         geom_point(size = 4) +
         coord_flip()+
         axis_looks+
         labs(x = " ", y = "VIP score")

#save plot
ggsave(
  filename = "cond_glucosinolates_VIP-scores.svg",
  plot = p_VIP,
  path = "./Plots/03_glucosinolates",
  scale = 1.5,
  width = 500,
  height = 700,
  units = "px",
  dpi = 300)


```

Now, let's analyze the data with a PERMANOVA. First, data is auto-scaled to calculate a distance matrix from which a PERMANOVA is run to see whether there is an effect of plant pathway induced, the type of induction and the interaction of these two factors. 

### PERMANOVA

```{r}

#PERMANOVA
UVmatrixdata<- santaR:::scaling_UV(matrixdata) #for auto-scaling using st. deviation
DistanceMatrix<-dist(UVmatrixdata, method="euclidean")

#make it a df again
MetaData <-as.data.frame(metadata) 
MetaData[,c(1:2)]<- lapply(MetaData[,c(1:2)], as.factor)

#run PERMANOVA
permanova <- adonis2(DistanceMatrix~Experiment*plant_pathway_induced,
                      data=MetaData, permutations = 999, method="bray")

#make table for Word
. <- c("Experiment (E)", "Plant pathway induced (PPI)", "E*PPI", "Residual", "Total") #names of factors
permanova_table <- cbind (.,as.data.frame(permanova))

table <- nice_table(permanova_table,
              title = c("Table X. PERMANOVA for Plant Glucosinolates", "Conditioning phase"),
              note = "999 number of permutations")

flextable::save_as_docx(table, path = "./Tables/03_glucosinolates/permanova_glucosinolates.docx")

#Run post-hoc tests

#Option 1
groups <- MetaData[["plant_pathway_induced"]]
mod <- betadisper(DistanceMatrix,groups)
permutest(mod)
plot(mod)
boxplot(mod)
mod.HSD <- TukeyHSD(mod)
mod.HSD
plot(mod.HSD)

#Option2
library(RVAideMemoire)

test.ppi <- pairwise.perm.manova(DistanceMatrix, MetaData$plant_pathway_induced,
                             test = c("Wilks"), nperm = 999, progress = TRUE,
                             p.method = "fdr", F = TRUE, R2 = TRUE)

test.exp <- pairwise.perm.manova(DistanceMatrix, MetaData$Experiment,
                             test = c("Wilks"), nperm = 999, progress = TRUE,
                             p.method = "fdr", F = T, R2 = T)

#Option 3
#Using this for now 
library(metagMisc)

#PPI
pairwise_ppi <- adonis_pairwise(x = MetaData, dd = DistanceMatrix,
                                group.var = "plant_pathway_induced", p.adj="fdr", permut=999)

pair_ppi_df <- as.data.frame(pairwise_ppi$Adonis.tab)
rownames(pair_ppi_df) <- pair_ppi_df$Comparison

#make table for Word -- Not working 
Comparisons <- c("Control-JA", "Control-SA", "JA-SA") #names of factors
pairwise_table <- cbind (Comparisons, as.data.frame(pair_ppi_df))

table <- nice_table(pair_ppi_df,
              title = c("Table X. PERMANOVA for Plant Phytohormones", "Conditioning phase"),
              note = "999 number of permutations")
str(pair_ppi_df)

#save it
table <- nice_table(pair_ppi_df,
              title = c("Table X. Post-Hoc PERMANOVA for Plant Phytohormones", "Conditioning phase"))
              
flextable::save_as_docx(table, path = "./Tables/02_phytohormones/posthoc_cond_ppi.docx")

#Experiment
pairwise_exp <- adonis_pairwise(x = MetaData, dd = DistanceMatrix,
                                group.var = "Experiment", p.adj="fdr", permut=999)

pair_exp_df <- as.data.frame(pairwise_exp$Adonis.tab) #no difference 


```



### Individual Barplots

Boxplot with rawdata points, colored by plant_pathway_induced and facet wraped by Experiment + colored in the upper boxes. For this, first I need to make a summary table df. 

```{r}
#summary stats 
df_gluco <- conditioning_long %>%
  group_by(Experiment, glucosinolate, plant_pathway_induced) %>%
  summarise(
    n = n(),
    mean = mean(concentration, na.rm = TRUE),
    sd = sd(concentration, na.rm = TRUE)) %>%
    mutate(se = sd/sqrt(n))

df_gluco

```

Now, let's plot it. I'd like to have a plot with those hormones with high concentrations and then a second plot with a smaller axis for those of low concentrations so differences can be better appreciated.

```{r}

big_axis <- c("Glu_total", "Glu_I3M", "Glu_1MOI3M", "Glu_3MSOP")
small_axis <- c("Glu_Allyl", "Glu_4MOI3M", "Glu_4MSOB")


# make a barplot for big-axis
p1 <- ggplot(data = subset(df_gluco, glucosinolate %in% big_axis),
            mapping = aes(x = glucosinolate, y = mean), colour= plant_pathway_induced)+
  geom_bar(aes(fill = plant_pathway_induced), stat = "identity",
           position = position_dodge(0.8), width = 0.8, color = "black")+
  geom_errorbar(aes(ymin = mean-se, ymax = mean+se, group = plant_pathway_induced),
                width = 0.2, position = position_dodge(0.8))+
  scale_fill_manual(values = color_pathway_induced)+
  geom_point(data = subset(conditioning_long, glucosinolate %in% big_axis),
              mapping = (aes(glucosinolate, concentration, fill = plant_pathway_induced)),
               shape = 21, size = 1.5, color = "black",
              position = position_jitterdodge(jitter.height = .1, jitter.width = .2))

p2 <- p1 + facet_wrap(~Experiment, labeller = as_labeller(Experiment_names)) +
              labs(y="(ng/g DW)", x=" ") +
              axis_looks +
              theme(panel.spacing = unit(0.2, "lines"))+
              theme(legend.position = "none")+
              scale_x_discrete(labels=c("Glu_total" = "Total", "Glu_I3M" = "I3M",
                                       "Glu_1MOI3M" = "1MOI3M", "Glu_3MSOP"="3MSOP"))


#save plot
ggsave(
  filename = "cond_glucosinolates_big.svg",
  plot = p2,
  path = "./Plots/03_glucosinolates",
  scale = 1.5,
  width = 1200,
  height = 700,
  units = "px",
  dpi = 300)


# make a barplot for small-axis
p1 <- ggplot(data = subset(df_gluco, glucosinolate %in% small_axis),
            mapping = aes(x = glucosinolate, y = mean), colour= plant_pathway_induced)+
  geom_bar(aes(fill = plant_pathway_induced), stat = "identity",
           position = position_dodge(0.8), width = 0.8, color = "black")+
  geom_errorbar(aes(ymin = mean-se, ymax = mean+se, group = plant_pathway_induced),
                width = 0.2, position = position_dodge(0.8))+
  scale_fill_manual(values = color_pathway_induced)+
  geom_point(data = subset(conditioning_long, glucosinolate %in% small_axis),
              mapping = (aes(glucosinolate, concentration, fill = plant_pathway_induced)),
               shape = 21, size = 1.5, color = "black",
              position = position_jitterdodge(jitter.height = .1, jitter.width = .2))

p2 <- p1 + facet_wrap(~Experiment) +
            theme(strip.background = element_blank(),
                  strip.text.x = element_blank())+ #no label on top
              labs(y="(ng/g DW)", x=" ") +
              axis_looks +
              theme(panel.spacing = unit(0.2, "lines"))+
              theme(legend.position = "none")+
              scale_x_discrete(labels=c("Glu_4MOI3M" = "4MOI3M", "Glu_4MSOB" = "4MSOB",
                                      "Glu_Allyl"="Sinigrin"))

#save plot
ggsave(
  filename = "cond_glucosinolate_small.svg",
  plot = p2,
  path = "./Plots/03_glucosinolates",
  scale = 1.5,
  width = 1200,
  height = 700,
  units = "px",
  dpi = 300)

#clean
rm(p1,p2)
dev.off()

```

These two plots will be fused in one in illustrator

## Feedback

###Plot indolic, aliphatic and total
First fix df 
```{r}

#these are my df's
gluco #full wide df
long #full long df 

#subset to feedback phase & keep only indolic and alyphatic for now 
feedback_long <- long %>% 
                 filter(glucosinolate %in% c("Glu_alyphatic","Glu_indolic")&
                 Phase == "feedback"&
                 soil_inocula != "non-conditioned")

levels(feedback_long$herbivory) #wrongly named

#change names of herbivory
feedback_long <- feedback_long %>% 
                 mutate(herbivory=case_when(
                   herbivory == "MeJA" ~ "caterpillar",
                   herbivory == "SA" ~ "aphid",
                   herbivory == "mock" ~ "not_infested",
                   herbivory == "without" ~ "not_infested",
                   TRUE ~ herbivory))
#make it a factor
feedback_long$herbivory <- factor(feedback_long$herbivory, levels=c("not_infested", "caterpillar", "aphid"))

#finally, delete not informative columns
feedback_long <- feedback_long %>% dplyr::select(-Stress, -plant_pathway_induced, -Glucosinolates, -treatment)%>%
                 relocate(soil_inocula, .after=Experiment)

#Now make a summary df
df_feedback_summ <- feedback_long  %>% 
                    group_by(Experiment, soil_inocula, herbivory, glucosinolate) %>%
                    summarise(
                    n = n(),
                    mean = mean(concentration, na.rm = TRUE),
                    sd = sd(concentration, na.rm = TRUE)) %>%
                    mutate(se = sd/sqrt(n))
```  

```{r}

#Finally, plot it by Experiment; First phyto
aliphatic <- ggplot(data = subset(df_feedback_summ, Experiment == "simulated" & glucosinolate == "Glu_alyphatic"),
       mapping = aes(x = glucosinolate, y = mean), colour= soil_inocula)+
  geom_bar(aes(fill = soil_inocula), stat = "identity",
           position = position_dodge(0.8), width = 0.8, color = "black")+
  geom_errorbar(aes(ymin = mean-se, ymax = mean+se, group = soil_inocula),
                width = 0.2, position = position_dodge(0.8))+
  geom_jitter(data = subset(feedback_long, Experiment == "simulated" & glucosinolate == "Glu_alyphatic"),
              mapping = (aes(glucosinolate, concentration,
                             fill = soil_inocula)),
              shape = 21, size = 1.5, alpha = 0.8, color = "black",
              position = position_jitterdodge(jitter.height = .1, jitter.width = .2))+
  scale_fill_manual(values = color_pathway_induced)+
  axis_looks+
  ylim(0,30)+
  theme(legend.position="none")+
  facet_wrap(~herbivory)+
  theme(panel.spacing = unit(0, "lines"))

indolic <- ggplot(data = subset(df_feedback_summ, Experiment == "simulated" & glucosinolate == "Glu_indolic"),
       mapping = aes(x = glucosinolate, y = mean), colour= soil_inocula)+
  geom_bar(aes(fill = soil_inocula), stat = "identity",
           position = position_dodge(0.8), width = 0.8, color = "black")+
  geom_errorbar(aes(ymin = mean-se, ymax = mean+se, group = soil_inocula),
                width = 0.2, position = position_dodge(0.8))+
  geom_jitter(data = subset(feedback_long, Experiment == "simulated" & glucosinolate == "Glu_indolic"),
              mapping = (aes(glucosinolate, concentration,
                             fill = soil_inocula)),
              shape = 21, size = 1.5, alpha = 0.8, color = "black",
              position = position_jitterdodge(jitter.height = .1, jitter.width = .2))+
  scale_fill_manual(values = color_pathway_induced)+
  axis_looks+
  ylim(0,60)+
  theme(legend.position="none")+
  facet_wrap(~herbivory)+
  theme(panel.spacing = unit(0, "lines"))

# Combine the two plots
library(patchwork)
combined_plot <- aliphatic / indolic

#save plot
mapply(function(x)
  ggsave(
  plot = last_plot(),
  filename =x,
  path = "./Plots/03_glucosinolates",
  scale = 1.8,
  width = 800,
  height = 700,
  units = "px",
  dpi = 300),
x=c("feedback_Phyto_barplot_aliphatic_indolic.svg",
    "feedback_Phyto_barplot_aliphatic_indolic.png"))

#Then insects
aliphatic_i <- ggplot(data = subset(df_feedback_summ, Experiment == "real_insects" & glucosinolate == "Glu_alyphatic"),
       mapping = aes(x = glucosinolate, y = mean), colour= soil_inocula)+
  geom_bar(aes(fill = soil_inocula), stat = "identity",
           position = position_dodge(0.8), width = 0.7, color = "black")+
  geom_errorbar(aes(ymin = mean-se, ymax = mean+se, group = soil_inocula),
                width = 0.2, position = position_dodge(0.8))+
  geom_jitter(data = subset(feedback_long, Experiment == "real_insects" & glucosinolate == "Glu_alyphatic"),
              mapping = (aes(glucosinolate, concentration,
                             fill = soil_inocula)),
              shape = 21, size = 1.5, alpha = 0.8, color = "black",
              position = position_jitterdodge(jitter.height = .1, jitter.width = .2))+
  scale_fill_manual(values = color_pathway_induced)+
  axis_looks+
  ylim(0,30)+
  theme(legend.position="none")+
  facet_wrap(~herbivory)+
  theme(panel.spacing = unit(0, "lines"))

indolic_i <- ggplot(data = subset(df_feedback_summ, Experiment == "real_insects" & glucosinolate == "Glu_indolic"),
       mapping = aes(x = glucosinolate, y = mean), colour= soil_inocula)+
  geom_bar(aes(fill = soil_inocula), stat = "identity",
           position = position_dodge(0.8), width = 0.7, color = "black")+
  geom_errorbar(aes(ymin = mean-se, ymax = mean+se, group = soil_inocula),
                width = 0.2, position = position_dodge(0.8))+
  geom_jitter(data = subset(feedback_long, Experiment == "real_insects" & glucosinolate == "Glu_indolic"),
              mapping = (aes(glucosinolate, concentration,
                             fill = soil_inocula)),
              shape = 21, size = 1.5, alpha = 0.8, color = "black",
              position = position_jitterdodge(jitter.height = .1, jitter.width = .2))+
  scale_fill_manual(values = color_pathway_induced)+
  axis_looks+
  ylim(0,60)+
  theme(legend.position="none")+
  facet_wrap(~herbivory)+
  theme(panel.spacing = unit(0, "lines"))

# Combine the two plots
combined_plot_i <- aliphatic_i / indolic_i

#save plot
mapply(function(x)
  ggsave(
  plot = last_plot(),
  filename =x,
  path = "./Plots/03_glucosinolates",
  scale = 1.8,
  width = 800,
  height = 700,
  units = "px",
  dpi = 300),
x=c("feedback_Insects_barplot_aliphatic_indolic.svg",
    "feedback_Insects_barplot_aliphatic_indolic.png"))

```


###Stats

Let's analyze the GLS by type independently in each experiment 

First for phyto
```{r}

#---- Phyto-Alyphatic ----
phyto <- subset(feedback_long, Experiment == "simulated" & glucosinolate == "Glu_alyphatic")
phyto %>% group_by(Experiment,soil_inocula,herbivory) %>% summarise(n=n()) #balanced

lm.p.a <- glm(log(concentration) ~ soil_inocula*herbivory,
              family=gaussian,
            data=phyto)

lm.p.a_table <- car::Anova(lm.p.a, type="II", test.statistic="LR")
simulateResiduals(lm.p.a, plot=T) #ok

#pairiwse differences
emmeans_herb <- emmeans(lm.p.a, pairwise~herbivory) #differences between NI and caterpillar
cld(emmeans_herb$emmeans, adjust="tukey")

#save result
factors <- c("Soil inoculum (S)", "Herbivory (H)", "S*H")
lm.p.a_table <- cbind (factors, as.data.frame(Anova(lm.p.a_table)))
lm.p.a_table$`Pr(>Chisq)` <- round(lm.p.a_table$`Pr(>Chisq)`, 3)
table <- nice_table(lm.p.a_table,
              title = c("Table X. GLM for aliphatic GLS in Feedback phase - Phyto"),
              note = "shoot_biomass_mg  ~ soil_inocula*herbivory, GLM-Gaussian")

table <- table %>%
         flextable::colformat_num(big.mark = "", digits = 3) 

flextable::save_as_docx(table, path = "./Tables/03_glucosinolates/feedback_aliGLS_phyto.docx")


#---- Phyto-Indolic ----
phyto_i <- subset(feedback_long, Experiment == "simulated" & glucosinolate == "Glu_indolic")
phyto_i %>% group_by(Experiment,soil_inocula,herbivory) %>% summarise(n=n()) #balanced

lm.p.i <- glm(log(concentration) ~ soil_inocula*herbivory,
              family=gaussian,
         data=phyto_i)

lm.p.i_table <- car::Anova(lm.p.i, type="II", test.statistic="LR")
simulateResiduals(lm.p.i, plot=T) #ok

#pairiwse differences
emmeans_soil <- emmeans(lm.p.i, pairwise~soil_inocula) #differences between Ctrl and SA

#save result
factors <- c("Soil inoculum (S)", "Herbivory (H)", "S*H")
lm.p.i_table <- cbind (factors, as.data.frame(lm.p.i_table))
lm.p.i_table$`Pr(>Chisq)` <- round(lm.p.i_table$`Pr(>Chisq)`, 3)
table <- nice_table(lm.p.i_table,
              title = c("Table X. GLM for indolic GLS in Feedback phase - Phyto"),
              note = "~ soil_inocula*herbivory")

table <- table %>%
         flextable::colformat_num(big.mark = "", digits = 3) 

flextable::save_as_docx(table, path = "./Tables/03_glucosinolates/feedback_indoGLS_phyto.docx")

```

For herbivore-infested experiment 
```{r}
#---- Insects-Alyphatic ----
insects <- subset(feedback_long, Experiment == "real_insects" & glucosinolate == "Glu_alyphatic")
insects %>% group_by(Experiment,soil_inocula,herbivory) %>% summarise(n=n()) #balanced

lm.i.a <- glm(log(concentration) ~ soil_inocula*herbivory,
              family=gaussian,
              data=insects)

lm.i.a_table <- car::Anova(lm.i.a, type="II", test.statistic="LR")
simulateResiduals(lm.i.a, plot=T) #ok

#save result
factors <- c("Soil inoculum (S)", "Herbivory (H)", "S*H")
lm.i.a_table <- cbind (factors, as.data.frame(lm.i.a_table))
lm.i.a_table$`Pr(>Chisq)` <- round(lm.i.a_table$`Pr(>Chisq)`, 3)
table <- nice_table(lm.i.a_table,
              title = c("Table X. GLM for aliphatic GLS in Feedback phase - Insects"),
              note = "shoot_biomass_mg  ~ soil_inocula*herbivory, GLM-Gaussian")

table <- table %>%
         flextable::colformat_num(big.mark = "", digits = 3) 

flextable::save_as_docx(table, path = "./Tables/03_glucosinolates/feedback_aliGLS_insects.docx")


#pairiwse differences
emmeans_soil <- emmeans(lm.i.a, pairwise~soil_inocula) #differences between CTR and JA and SA
emmeans_herb <- emmeans(lm.i.a, pairwise~herbivory)


#---- Insects-Indolic ----
insects_i <- subset(feedback_long, Experiment == "real_insects" & glucosinolate == "Glu_indolic")
insects_i %>% group_by(Experiment,soil_inocula,herbivory) %>% summarise(n=n()) #balanced

lm.i.i <- glm(log(concentration) ~ soil_inocula*herbivory,
              family=gaussian,
              data=insects_i)

lm.i.i_table <- car::Anova(lm.i.i, type="II", test.statistic="LR")
simulateResiduals(lm, plot=T) #ok

#pairiwse differences
emmeans_herb <- emmeans(lm.i.i, pairwise~herbivory) #differences between Ctrl and SA

#save result
factors <- c("Soil inoculum (S)", "Herbivory (H)", "S*H")
lm.i.i_table <- cbind (factors, as.data.frame(lm.i.i_table))
lm.i.i_table$`Pr(>Chisq)` <- round(lm.i.i_table$`Pr(>Chisq)`, 3)
table <- nice_table(lm.i.i_table,
              title = c("Table X. GLM for indolic GLS in Feedback phase - Insects"),
              note = "~ soil_inocula*herbivory")

table <- table %>%
         flextable::colformat_num(big.mark = "", digits = 3) 

flextable::save_as_docx(table, path = "./Tables/03_glucosinolates/feedback_indoGLS_insects.docx")


```


```{r}
save.image("./R_Environments/glucosinolates/glucosinolates.Rdata")
```




```{r}

#subset existing df to only feedback
gluco
long

#subsetting df to only conditioning
feedback <- gluco %>% 
                filter(Phase == "feedback")%>%
                subset(soil_inocula != "non-conditioned")%>% 
                droplevels()

#subsetting df to only feedback
feedback_long <- long %>% 
                filter(Phase == "feedback")%>%
                subset(soil_inocula != "non-conditioned")%>%
                droplevels()

#change names of soil_inocula
levels(feedback$soil_inocula)
feedback <- feedback  %>%
            mutate(soil_inocula=factor(soil_inocula, labels=c("control_inocula", 
                                                              "JA_inocula",
                                                              "SA_inocula")))
levels(feedback_long$soil_inocula)
feedback_long <- feedback_long  %>%
                  mutate(soil_inocula=factor(soil_inocula, labels=c("control_inocula", 
                                                              "JA_inocula",
                                                              "SA_inocula")))
#change names of herbivory
levels(feedback$herbivory)
feedback <- feedback  %>%
            mutate(herbivory=factor(herbivory, levels=c("without", "caterpillar", "aphid")))

feedback <- feedback  %>%
            mutate(herbivory=factor(herbivory,
                                    labels=c("uninfested", "Mamestra", "Myzus")))

levels(feedback_long$herbivory)
feedback_long <- feedback_long  %>%
                  mutate(herbivory=factor(herbivory, levels=c("without", "caterpillar", "aphid")))

feedback_long <- feedback_long  %>%
                mutate(herbivory=factor(herbivory,
                                    labels=c("uninfested", "Mamestra", "Myzus")))

```

### PLSDA

```{r}

#The following code was written by Els van de Zande, thank you Els!

View(feedback)

str(feedback)
matrixdata<-feedback[,11:18]# keep only values
matrixdata <- as.matrix(matrixdata[,])# make it numeric
metadata<-feedback[,c("Experiment", "soil_inocula", "herbivory")]
metadata_number <- metadata %>% #change from characters to numbers
                     mutate(Experiment = case_when(Experiment == 'simulated' ~ '1',
                                                   TRUE ~ '2')) %>%
                     mutate(soil_inocula = case_when(soil_inocula == 'control_inocula' ~ '1',
                                                     soil_inocula == 'JA_inocula' ~ '2',
                                                       TRUE ~ '3'))%>% #SA inocula
                     mutate(herbivory = case_when(herbivory == 'uninfested' ~ '1',
                                                  herbivory == 'Mamestra' ~ '2',
                                                  TRUE ~ '3')) #Myzus
View(metadata_number)
MetaData <-data.matrix(metadata_number) #makes a numeric matrix

#PLS-DA
plsda <-opls(matrixdata,MetaData, predI=2)
plsda_center <-opls(matrixdata,MetaData, scaleC = "center", predI=2)
plsda_pareto <-opls(matrixdata,MetaData, scaleC = "pareto", predI=2)#removes the relative importance of large values
                                                                    #but keeps the data structure intact.
#check
plsda_pareto@summaryDF
plot(plsda_pareto)
plsda_pareto@scoreMN
plot(plsda_pareto,
     typeVc = "x-score",
     parAsColFcVn = MetaData[,"herbivory"],
     parLabVc = as.character(MetaData[,"herbivory"]))

#split by factors
soil_inocula_Factor <- MetaData[,"soil_inocula"]
experiment_Factor <- MetaData[,"Experiment"]
herbivory_Factor <- MetaData[,"herbivory"]

plsda_pareto2 <-opls(matrixdata,soil_inocula_Factor, scaleC = "pareto", permI=1000) #plant pathway induced
plsda_pareto3 <-opls(matrixdata,experiment_Factor, scaleC = "pareto", permI=1000) #Experiment 
plsda_pareto4 <-opls(matrixdata,herbivory_Factor, scaleC = "pareto", permI=1000) #Herbivory, something goes wrong

#get scores to plot them
scores<-plsda_pareto@scoreMN
scores_df<-as.data.frame(scores) #make new df 

scores_df$p1 <- as.numeric(scores_df$p1)
scores_df$p2 <- as.numeric(scores_df$p2)

scores_df2 <- scores_df %>% #add info
              mutate(Experiment = as.numeric(metadata_number$Experiment))%>%
              mutate(soil_inocula = as.numeric(metadata_number$soil_inocula))%>%
              mutate(herbivory = as.numeric(metadata_number$herbivory))
              
#Plot the scores using ggplot2
scores_df2[c(3:5)] <- lapply(scores_df2[c(3:5)] , factor)
str(scores_df2)


scoreplot<-ggplot(scores_df2, aes(x = p1, y = p2,
                                  fill=soil_inocula,#inside
                                  shape=herbivory, #shape
                                  color=Experiment))+ #outline
  geom_point(size = 4, stroke=0.5, alpha=0.8,
             aes(color=Experiment, fill=soil_inocula))+
  scale_shape_manual(values=c(21,22,24))+
  scale_color_manual(values = c("black", "white"))+ #outline
  scale_fill_manual(values=color_pathway_induced)+ #inside
  labs(x = "t1(36%)", y = "t2(34%)",
  color="Experiment", shape="herbivory", fill="Soil inocula")+
  axis_looks

  
#now add loadings
loadings<-plsda_pareto@loadingMN
loadings_df<-as.data.frame(loadings)

loadings_df$p1 <- as.numeric(loadings_df$p1)
loadings_df$p2 <- as.numeric(loadings_df$p2)

#makes plot with arrows marking direction
rownames(loadings)
glucosinolate_names <- c("Total","3MSOP","4MSOB","Sinigrin","4OHI3M","I3M","4MOI3M","1MOI3M")

Biplot <-scoreplot+
         geom_hline(yintercept=0, linetype='dotted', col = 'gray20',linewidth = 0.2 )+
         geom_vline(xintercept=0, linetype='dotted', col = 'gray20', linewidth = 0.2)+
         geom_segment(data = loadings_df, aes(x = 0, y = 0, xend = (p1*20), yend = (p2*20)),
                      arrow = arrow(length = unit(0.5, "picas"), type="closed"),
                      colour = "black", linewidth = 0.5, inherit.aes = FALSE)+
         geom_text(aes(x = (p1*20), y = (p2*20), label = glucosinolate_names,
                angle = 0.45, hjust = -.1, vjust = .2),
                 data = loadings_df,
                 colour = "black", size=4, inherit.aes = FALSE)+
        #ylim(-60,60)+
        #xlim(-100,100)+
        axis_looks
         #theme(legend.position = "none")

#save plot
ggsave(
  filename = "feedback_glucosinolates_PLSDA.svg",
  plot = Biplot,
  path = "./Plots/03_glucosinolates",
  scale = 1.5,
  width = 1200,
  height = 700,
  units = "px",
  dpi = 300)


#make same plot but colored by herbivory
scoreplot<-ggplot(scores_df2, aes(x = p1, y = p2,
                                  fill=herbivory,#inside
                                  shape=soil_inocula, #shape
                                  color=Experiment))+ #outline
  geom_point(size = 4, stroke=0.5, alpha=0.8,
             aes(color=Experiment, fill=herbivory))+
  scale_shape_manual(values=c(21,22,24))+
  scale_color_manual(values = c("black", "white"))+ #outline
  scale_fill_manual(values=color_pathway_induced)+ #inside
  labs(x = "t1(36%)", y = "t2(34%)",
  color="Experiment", shape="Soil inocula", fill="Herbivory")+
  axis_looks

Biplot <-scoreplot+
         geom_hline(yintercept=0, linetype='dotted', col = 'gray20',linewidth = 0.2 )+
         geom_vline(xintercept=0, linetype='dotted', col = 'gray20', linewidth = 0.2)+
         geom_segment(data = loadings_df, aes(x = 0, y = 0, xend = (p1*20), yend = (p2*20)),
                      arrow = arrow(length = unit(0.5, "picas"), type="closed"),
                      colour = "black", linewidth = 0.5, inherit.aes = FALSE)+
         geom_text(aes(x = (p1*20), y = (p2*20), label = glucosinolate_names,
                angle = 0.45, hjust = -.1, vjust = .2),
                 data = loadings_df,
                 colour = "black", size=4, inherit.aes = FALSE)+
        #ylim(-60,60)+
        #xlim(-100,100)+
        axis_looks
         #theme(legend.position = "none")

#save plot
ggsave(
  filename = "feedback_glucosinolates_PLSDA_ByHerbivory.svg",
  plot = Biplot,
  path = "./Plots/03_glucosinolates",
  scale = 1.5,
  width = 1200,
  height = 700,
  units = "px",
  dpi = 300)


#Let's also get the VIP scores 

#extract VIP (Variable's importance) scores and CV-value
VIPmatrixdataPLSDA<-getVipVn(plsda_pareto, orthoL = FALSE)

VIP <- as.data.frame(VIPmatrixdataPLSDA)
VIP <- VIP %>%
        arrange(VIP)%>%
        rename(VIP = VIPmatrixdataPLSDA)%>%
        mutate(glucosinolate=rownames(.))%>%
        mutate(glucosinolate)

VIP$glucosinolate <- as.factor(VIP$glucosinolate)

#plot it to have it next to the plot
p_VIP <- VIP %>%
         mutate(glucosinolate=factor(glucosinolate, levels=glucosinolate)) %>%
         ggplot(aes(x = glucosinolate, y = VIP)) +
         geom_point(size = 4) +
         coord_flip()+
         axis_looks+
         labs(x = " ", y = "VIP score")

#save plot
ggsave(
  filename = "feedback_glucosinolates_VIP-scores.svg",
  plot = p_VIP,
  path = "./Plots/03_glucosinolates",
  scale = 1.5,
  width = 500,
  height = 700,
  units = "px",
  dpi = 300)


```

Now, let's analyze the data with a PERMANOVA. First, data is auto-scaled to calculate a distance matrix from which a PERMANOVA is run to see whether there is an effect of plant pathway induced, the type of induction and the interaction of these two factors. 

### PERMANOVA

```{r}

#PERMANOVA
UVmatrixdata<- santaR:::scaling_UV(matrixdata) #for auto-scaling using st. deviation
DistanceMatrix<-dist(UVmatrixdata, method="euclidean")

#make it a df again
MetaData <-as.data.frame(metadata) 
MetaData[,c(1:3)]<- lapply(MetaData[,c(1:3)], as.factor)

#run PERMANOVA
permanova <- adonis2(DistanceMatrix~soil_inocula*herbivory*Experiment,
                      data=MetaData, permutations = 999, method="bray")

#make table for Word
. <- c("Soil inocula (S)", "Herbivory (H)", "Experiment (E)", "S*H", "S*E", "H*E", "S*H*E",
       "Residual", "Total") #names of factors
permanova_table <- cbind (.,as.data.frame(permanova))

table <- nice_table(permanova_table,
              title = c("Table X. PERMANOVA for Glucosinolates", "Feedback phase",
                        "DistanceMatrix~soil_inocula*herbivory*Experiment"),
              note = "999 number of permutations")

flextable::save_as_docx(table, path = "./Tables/03_glucosinolates/permanova_glucosinolates_feedback.docx")

#Run post-hoc tests

#Option 1
groups <- MetaData[["herbivory"]]
mod <- betadisper(DistanceMatrix,groups) #checking beta-dispersion
permutest(mod)
plot(mod)
boxplot(mod)
mod.HSD <- TukeyHSD(mod)
mod.HSD
plot(mod.HSD)

#Option 2
#Using this for now 
library(metagMisc)

#Soil inocula
pairwise_ppi <- adonis_pairwise(x = MetaData, dd = DistanceMatrix,
                                group.var = "herbivory", p.adj="fdr", permut=999)

pair_ppi_df <- as.data.frame(pairwise_ppi$Adonis.tab)
rownames(pair_ppi_df) <- pair_ppi_df$Comparison

#make table for Word -- Not working 
Comparisons <- c("Control-inocula vs JA-inocula", "Control-inocula vs SA-inocula",
                 "JA-inocula vs SA-inocula") #names of factors

pairwise_table <- cbind (Comparisons, as.data.frame(pair_ppi_df))

table <- nice_table(pairwise_table, #Not working
              title = c("Table X. Post-hoc for Soil Inocula", "Feedback phase"),
              note = "999 number of permutations")

str(pair_ppi_df)

#save it
table <- nice_table(pair_ppi_df, #Not working
              title = c("Table X. Post-Hoc PERMANOVA for Plant Phytohormones", "Conditioning phase"))
              
#flextable::save_as_docx(table, path = "./Tables/02_phytohormones/posthoc_cond_ppi.docx")

```

From: https://www.researchgate.net/post/How_should_I_correctly_manage_PERMANOVA_for_factors_with_interactions

A significant Permanova means one of three things. 1) There is a difference in the location of the samples (i.e. the average community composition), 2) There is a difference in the dispersion of the samples (i.e. the variability in the community composition), or 3) There is a difference in both the location and the dispersion.
So, if you get a significant Permanova you'll want to distinguish between the three options. That of course is why you need to run the permdisp. If you get a non-significant Permdisp you can conclude the first option above is the correct one. If you get a significant Permdisp then it is either the second or third option (that is there is definitely a difference in dispersion and maybe a difference in location). There is no fool-proof way to distinguish between these two, but looking at an MDS plot of the data will hopefully help you do so (also see note 2 below).
A few notes.
1) Permanova is not as powerful as permdisp at detecting differences in dispersion so it's possible to get a non-significant Permanova and a significant permdisp. This would mean that you have a difference in dispersion only. That can be an important ecologically in itself.
2) Transforming your data (square-root, cubed root, log, presence-absence) can reduce your dispersion and is a potential way to help distinguish between dispersion and both dispersion and location. However you need to take the transformation into account when interpreting the results. A significant permanova on raw data and a significant permanova on presence-absence data are interpreted differently, but that is another discussion in itself.
