==> 04_insect_performance.Rmd <==
---
title: "04_insect_performance"
author: "Marcela Aragon"
date: "16/1/2024"
output: 
  html_document:
    toc: true
    toc_depth: 3
    toc_float: 
     collapsed: false
     smooth_scroll: true
    df_print: paged
    code_folding: hide
    keep_md: true
editor_options: 
  chunk_output_type: console
---

This script will deal with ploting and analyzing the concentration of X different phytohormones in both conditioning and feedback phases;

# General settings 


## Loading packages

```{r loading packages, echo=TRUE}

#Set working directory to project directory 
setwd("./")
getwd() #ok

#data
library(knitr)
library(dplyr)
library(tibble)

#plotting
library(ggplot2)
library(patchwork)
library(rempsyc) #tables

#stats
library(lme4)
library(DHARMa)
library(lmtest)
library(fitdistrplus)
library(AICcmodavg)
library(emmeans)
library(multcompView)
library(multcomp)
library(postHoc)
library(car)

```

## Global settings for plots 

```{r settings for plots}

# Setting up white background for plots
theme_set(theme_bw())

#Set colors
color_stress <- c("#BFBFBF", "#E69F00", "#4472C4")
color_experiment <- c("#CC3399", "#70AD47")

color_ctrl_vs_JA <- c("#BFBFBF", "#E69F00")
color_ctrl_vs_SA <- c("#BFBFBF",  "#4472C4")
color_control <- "#BFBFBF"
color_JA <- "#E69F00"
color_SA <- "#4472C4"
color_experiment <- c("#CC3399", "#70AD47")
fonts <- windowsFonts()

color_pathway_induced <- c(color_control, color_JA, color_SA) #control, JA and SA

# Set default axis looks
axis_looks <- theme(axis.text.x = element_text(colour = "black", size = 9,
                                               face = "bold", angle=0, hjust=0.5))+ #text in X
  theme(axis.text.y = element_text(colour = "black", size = 9, face = "bold"))+ #text in y
  theme(axis.title=element_text(size=11, face = "bold"))+ #all texts
  theme(axis.title.y = element_text(margin = ggplot2::margin(t = 0, r = 10, b = 0, l = 0)))+
  theme(panel.grid.major = element_blank(), #no grid
        panel.grid.minor = element_blank()) #no grid
  #theme(legend.position="none")+ #no legend


#Get names for boxes in facet wrap
Experiment_names <- c(
                    `simulated` = "Simulated Herbivory inocula",
                    `real_insects` = "Insect Herbivory inocula")

```

# Caterpillars

```{r}
#load data
caterpillars <- readRDS("./plant_data/caterpillars.rds")

#explore 
summary(caterpillars)
colnames(caterpillars)

#change levels of Experiment
levels(caterpillars$Experiment)
levels(caterpillars$soil_inocula)
caterpillars <- caterpillars %>%
                mutate(Experiment = factor(Experiment,
                                           levels = c("simulated", "real_insects" )))%>%
                mutate(soil_inocula = factor(soil_inocula,
                                                      levels=c("non-conditioned", "control-conditioned",
                                                               "JA-conditioned", "SA-conditioned")))

#let's remove non-conditioned for now
caterpillars <-  caterpillars %>% 
                   subset(soil_inocula != "non-conditioned")%>%
                   subset(caterpillar_number != 0)%>% #removed 3 plants with 0 caterpillars 
                   droplevels()





#make it long
#making it long 
cat_long <- caterpillars %>% 
               tidyr::pivot_longer(cols = c(7:11), 
               names_to = "caterpillar",
               values_to = "weight",
               values_drop_na = TRUE)

#make summary
cat_summary <- cat_long %>% 
            group_by(Experiment, soil_inocula, ID) %>%
  summarise(
    n = n(),
    mean = mean(weight, na.rm = TRUE),
    sd = sd(weight, na.rm = TRUE)) %>%
    mutate(se = sd/sqrt(n))

```


## Plot
```{r}

#check correlation between caterpillar number and weight
plot(cat_long$caterpillar_number,cat_long$weight) #it's the opposite of what I expected 

#now check by soil inocula and Experiment
ggplot(cat_summary, mapping=aes(x=n, y=mean, group=soil_inocula))+
  geom_point()+
  geom_smooth(method="lm", se=TRUE)+
  facet_wrap(Experiment~soil_inocula) #difficult to tell, it's not really balanced

#plot boxplot Phyto
phyto <- ggplot(data = subset(cat_summary, Experiment == "simulated"),
       mapping = aes(x = soil_inocula, y = mean), colour= soil_inocula)+
  geom_boxplot(mapping = aes(fill= soil_inocula), position = position_dodge(width = 0.8), lwd=0.8)+
  geom_jitter(data = subset(cat_summary, Experiment == "simulated"),
              mapping = (aes(soil_inocula, mean, fill = soil_inocula)),
              shape = 21, size = 1.8, color = "black",
              position = position_jitterdodge(jitter.height = .1, jitter.width = .2))+
  scale_fill_manual(values = color_pathway_induced)+
  ylim(0,35)+
  axis_looks+
  theme(legend.position = "none")+
  labs(y="M. brassicae weight (mg) 8dpi", x=" ")

#plot boxplot Insects
insects <- ggplot(data = subset(cat_summary, Experiment == "real_insects"),
       mapping = aes(x = soil_inocula, y = mean), colour= soil_inocula)+
  geom_boxplot(mapping = aes(fill= soil_inocula), position = position_dodge(width = 0.8), lwd=0.8)+
  geom_jitter(data = subset(cat_summary, Experiment == "real_insects"),
              mapping = (aes(soil_inocula, mean, fill = soil_inocula)),
              shape = 21, size = 1.8, color = "black",
              position = position_jitterdodge(jitter.height = .1, jitter.width = .2))+
  scale_fill_manual(values = color_pathway_induced)+
  axis_looks+
  ylim(0,35)+
  theme(legend.position = "none")+
  labs(y="M. brassicae weight (mg) 8dpi", x=" ")

#combine plots
library(patchwork)
combined_plot <- phyto/insects

#save plot
mapply(function(x)
  ggsave(
  plot = last_plot(),
  filename =x,
  path = "./Plots/04_insect_performance",
  scale = 1.8,
  width = 500,
  height = 700,
  units = "px",
  dpi = 300),
x=c("caterpillar_weight.svg",
    "caterpillar_weight.png"))
```

## Stats

```{r}

#Linear Model Phyto
phyto <- subset(cat_summary, Experiment == "simulated")
phyto %>% group_by(Experiment,soil_inocula) %>% summarise(n=n()) #not-balanced

#lm
lm.p <- glm(mean ~ soil_inocula, 
            family=gaussian,
         data=phyto)

lm.p_table <- car::Anova(lm.p, type="II", test.statistic="LR")
simulateResiduals(lm.p, plot=T) #ok

#save result
factors <- c("Soil inocula(S)")
lm.p_table <- cbind (factors, as.data.frame(lm.p_table))
lm.p_table$`Pr(>Chisq)` <- round(lm.p_table$`Pr(>Chisq)`, 3)
table <- nice_table(lm.p_table,
              title = c("Table X. GLM for aliphatic GLS in Feedback phase - Insects"),
              note = "~ soil_inocula*herbivory")

table <- table %>%
         flextable::colformat_num(big.mark = "", digits = 3) 

flextable::save_as_docx(table, path = "./Tables/04_insect_performance/caterpillar_weight_phyto.docx")

#Linear Model Insects
insects <- subset(cat_summary, Experiment == "real_insects")
insects %>% group_by(Experiment,soil_inocula) %>% summarise(n=n()) #balanced


#glmer
lm.i <- glm(mean ~ soil_inocula, 
            family=gaussian,
         data=insects)

lm.i_table <- car::Anova(lm.i, type="II", test.statistic="LR")
simulateResiduals(lm.i, plot=T) #ok

#save result
factors <- c("Soil inocula(S)")
lm.i_table <- cbind (factors, as.data.frame(lm.i_table))
lm.i_table$`Pr(>Chisq)` <- round(lm.i_table$`Pr(>Chisq)`, 3)
table <- nice_table(lm.i_table,
              title = c("Table X. GLM for aliphatic GLS in Feedback phase - Insects"),
              note = "~ soil_inocula*herbivory")

table <- table %>%
         flextable::colformat_num(big.mark = "", digits = 3) 

flextable::save_as_docx(table, path = "./Tables/04_insect_performance/caterpillar_weight_insects.docx")

```

```{r}
#difference between experiments

exp_summ <- cat_summary %>% 
            as.data.frame() %>% 
            group_by(Experiment, soil_inocula) %>% 
            summarise(mean=mean, na.rm=TRUE)


write.csv(cat_summary, file="./cat_summary.csv")

```

*Test later with glmer using plant as random factor, now my lme4 is not working
```{r}

phyto_long <- subset(cat_long, Experiment == "simulated")
saveRDS(phyto_long, "./phyto_long.rdata")
write.csv(phyto_long, "phyto_long.csv", row.names = FALSE)

#Both
#glme <- glmer(weight ~ soil_inocula*Experiment+ (1|ID),
#              family=Gamma, data=cat_long)


#Get model info 
library(jtools)
summ(glme)
Anova(glme)
summary(glme)

#check
library("blmeco") 
dispersion_glmer(glme) #it shouldn't be over 1.4 so it's ok
qqnorm(resid(glme), main="normal qq-plot, residuals") #good
qqline(resid(glme))


par(mfrow=c(2,2))
qqnorm(resid(glme), main="normal qq-plot, residuals")
qqline(resid(glme))

plot(fitted(glme), resid(glme)) #residuals vs fitted
abline(h=0)

#save table
. <- c("Soil inocula (S)", "Experiment (E)", "S*E")
glm_table <- cbind (.,as.data.frame(Anova(glme)))
table <- nice_table(glm_table,
              title = c("Table X. Generalized Linear Model for Caterpillar Weight"),
              note = "Gamma distribution")

flextable::save_as_docx(table, path = "./Tables/04_insect_performance/caterpillar_weight_glm.docx")

```

# Aphids

```{r}
#load data
aphids <- readRDS("./plant_data/aphids.rds")

#explore 
summary(aphids)
colnames(aphids)

#change levels of Experiment
levels(aphids$Experiment)
levels(aphids$soil_inocula)
aphids <- aphids %>%
                mutate(Experiment = factor(Experiment,
                                           levels = c("simulated", "real_insects" )))%>%
                mutate(soil_inocula = factor(soil_inocula,
                                                      levels=c("non-conditioned", "control-conditioned",
                                                               "JA-conditioned", "SA-conditioned")))

#let's remove non-conditioned for now
aphids <-  aphids %>% 
                   subset(soil_inocula != "non-conditioned")%>%
                   subset(ID != 145 & ID!= 351)%>% #removed 2 plants with wrong infestation 
                   droplevels()

#make it long
aphids_long <- aphids %>% 
               tidyr::pivot_longer(cols = c(7:8), 
               names_to = "time_point",
               values_to = "count",
               values_drop_na = TRUE)

str(aphids_long)
aphids_long[c(1:2,7)] <- lapply(aphids_long[c(1:2,7)], as.factor)

aphids_long$time_point <- factor(aphids_long$time_point,
                                 levels = c("dpi_8", "dpi_14")) 

```



## Plot
```{r}


#Phyto
aphids_8dpi <- ggplot(data = subset(aphids_long, Experiment == "simulated" & time_point == "dpi_8"),
                 mapping = aes(x = soil_inocula, y = count), 
              colour= soil_inocula)+
  geom_boxplot(mapping = aes(fill= soil_inocula), position = position_dodge(width = 0.8), lwd=0.8)+
  geom_jitter(data = subset(aphids_long, Experiment == "simulated" & time_point == "dpi_8"),
              mapping = (aes(soil_inocula, count, fill = soil_inocula)),
              shape = 21, size = 1.8, alpha=0.9, color = "black",
              position = position_jitterdodge(jitter.height = .1, jitter.width = .2))+
  scale_fill_manual(values = color_pathway_induced)+
  axis_looks+
  ylim(0,80)+
  #facet_wrap(~time_point, scales="free_y")+
  theme(legend.position="none")

aphids_14dpi <- ggplot(data = subset(aphids_long, Experiment == "simulated" & time_point == "dpi_14"),
                 mapping = aes(x = soil_inocula, y = count), 
              colour= soil_inocula)+
  geom_boxplot(mapping = aes(fill= soil_inocula), position = position_dodge(width = 0.8), lwd=0.8)+
  geom_jitter(data = subset(aphids_long, Experiment == "simulated" & time_point == "dpi_14"),
              mapping = (aes(soil_inocula, count, fill = soil_inocula)),
              shape = 21, size = 1.8, alpha=0.9,color = "black",
              position = position_jitterdodge(jitter.height = .1, jitter.width = .2))+
  scale_fill_manual(values = color_pathway_induced)+
  axis_looks+
  ylim(0,600)+
  #facet_wrap(~time_point, scales="free_y")+
  theme(legend.position="none")


#insects
aphids_8dpi_i <- ggplot(data = subset(aphids_long, Experiment == "real_insects" & time_point == "dpi_8"),
                 mapping = aes(x = soil_inocula, y = count), 
              colour= soil_inocula)+
  geom_boxplot(mapping = aes(fill= soil_inocula), position = position_dodge(width = 0.8), lwd=0.8)+
  geom_jitter(data = subset(aphids_long, Experiment == "real_insects" & time_point == "dpi_8"),
              mapping = (aes(soil_inocula, count, fill = soil_inocula)),
              shape = 21, size = 1.8, alpha=0.9,color = "black",
              position = position_jitterdodge(jitter.height = .1, jitter.width = .2))+
  scale_fill_manual(values = color_pathway_induced)+
  axis_looks+
  ylim(0,80)+
  #facet_wrap(~time_point, scales="free_y")+
  theme(legend.position="none")

aphids_14dpi_i <- ggplot(data = subset(aphids_long, Experiment == "real_insects" & time_point == "dpi_14"),
                 mapping = aes(x = soil_inocula, y = count), 
              colour= soil_inocula)+
  geom_boxplot(mapping = aes(fill= soil_inocula), position = position_dodge(width = 0.8), lwd=0.8)+
  geom_jitter(data = subset(aphids_long, Experiment == "real_insects" & time_point == "dpi_14"),
              mapping = (aes(soil_inocula, count, fill = soil_inocula)),
              shape = 21, size = 1.8, alpha=0.9,color = "black",
              position = position_jitterdodge(jitter.height = .1, jitter.width = .2))+
  scale_fill_manual(values = color_pathway_induced)+
  axis_looks+
  ylim(0,600)+
  #facet_wrap(~time_point, scales="free_y")+
  theme(legend.position="none")

#combined plot
library(patchwork)
aphids <- (aphids_8dpi+aphids_14dpi)/(aphids_8dpi_i+aphids_14dpi_i)


#save plot
mapply(function(x)
  ggsave(
  plot = last_plot(),
  filename =x,
  path = "./Plots/04_insect_performance",
  scale = 1.8,
  width = 1000,
  height = 700,
  units = "px",
  dpi = 300),
x=c("aphid_count_bothTimepoints.svg",
    "aphid_count_bothTimepoints.png"))
```


Two timepoints
```{r}

#First, make a summary table
#summary stats 
df_aphids <- aphids_long %>%
  group_by(Experiment, soil_inocula, time_point) %>%
  summarise(
    n = n(),
    mean = mean(count, na.rm = TRUE),
    sd = sd(count, na.rm = TRUE)) %>%
    mutate(se = sd/sqrt(n))

df_aphids

#plot it
p <- ggplot(data = df_aphids, mapping = aes(x = time_point, y = mean, 
              colour= soil_inocula, group=soil_inocula))+
  geom_point(size=3)+
  geom_line(aes(group = soil_inocula), size = 0.6, linetype="solid") +
  geom_errorbar(aes(ymin = mean - se, ymax = mean + se), width = 0.15, size=0.8) +  # Error bars
  facet_wrap(~ Experiment)+
  scale_colour_manual(values = color_pathway_induced)


ggsave(
  filename = "aphid_number_two_timepoints_lines.svg",
  plot = p,
  path = "./Plots/04_insect_performance",
  scale = 1.5,
  width = 1200,
  height = 700,
  units = "px",
  dpi = 300)

```


## Stats

```{r}

#full df
aphids_long

#subset to phyto
phyto <- subset(aphids_long, Experiment == "simulated")
phyto %>% group_by(Experiment,soil_inocula) %>% summarise(n=n()) #balanced

glm.p <- glm(count ~ soil_inocula*time_point,
              family=Gamma(link="log"),
              data=phyto)

#check
simulateResiduals(glm.p , plot=T) #ok


#get summary
x <- car::Anova(glm.p , type="II", test.statistic="LR")
x <- as.data.frame(x)

#save result
factors <- c("Soil inocula (S)", "Time point(T)", "S*T")
glm.p_table <- cbind (factors, x)
glm.p_table$`Pr(>Chisq)` <- round(glm.p_table$`Pr(>Chisq)`, 3)
table <- nice_table(glm.p_table,
              title = c("Table X. GLM for aphid number over time - Phyto"),
              note = "~ soil_inocula*Timepoint")

table <- table %>%
         flextable::colformat_num(big.mark = "", digits = 3) 

flextable::save_as_docx(table, path = "./Tables/04_insect_performance/aphid_number_phyto.docx")


#---- insects ----
insects <- subset(aphids_long, Experiment == "real_insects")
insects %>% group_by(Experiment,soil_inocula) %>% summarise(n=n()) #not-balanced

glm.i <- glm(count ~ soil_inocula*time_point,
              family=Gamma(link="log"),
              data=insects)

#check
simulateResiduals(glm.i, plot=T) #ok

#get summary
y <- car::Anova(glm.i , type="II", test.statistic="LR")
y <- as.data.frame(y)

#save result
factors <- c("Soil inoculum (S)", "Time point(T)", "S*T")
glm.i_table <- cbind (factors, y)
glm.i_table$`Pr(>Chisq)` <- round(glm.i_table$`Pr(>Chisq)`, 3)
table <- nice_table(glm.i_table,
              title = c("Table X. GLM for aphid number over time - Insects"),
              note = "~ soil_inocula*Timepoint")

table <- table %>%
         flextable::colformat_num(big.mark = "", digits = 3) 

flextable::save_as_docx(table, path = "./Tables/04_insect_performance/aphid_number_insects.docx")


```

#save
```{r}
save.image("./R_Environments/insect_performance/insect_performance.Rdata")
```
