==> 05_EPG_and_AphidPerformance_PSF.Rmd <==
# Loading packages

```{r loading packages, echo=TRUE}

#Set working directory to project directory 
setwd("./")
getwd() #ok

#data
library(knitr)
library(dplyr)
library(tibble)
library(readxl)

#plotting
library(ggplot2)
library(ggpubr)
library(rempsyc) #tables

#analysis
library(DHARMa)
library(pscl)
library(nlme)
library(lme4)
library(lmerTest)
library(plotrix) #stderror

```

#settings for plots
```{r}

theme_set(theme_bw())
color_stress <- c("#BFBFBF", "#319B67")

# Set default axis looks
axis_looks <- theme(axis.text.x = element_text(colour = "black", size = 9,
                                               face = "bold", angle=0, hjust=0.5))+ #text in X
  theme(axis.text.y = element_text(colour = "black", size = 9, face = "bold"))+ #text in y
  theme(axis.title=element_text(size=11, face = "bold"))+ #all texts
  theme(axis.title.y = element_text(margin = ggplot2::margin(t = 0, r = 10, b = 0, l = 0)))+
  theme(panel.grid.major = element_blank(), #no grid
        panel.grid.minor = element_blank()) #no grid
```

#Loading data

```{r}

#aphid population and plant performance
aphid <- read_xlsx("./plant_data/raw_data/EPG/Aphid_population_adjusted.xlsx")
plant <- read_xlsx("./plant_data/raw_data/EPG/Plant performance.xlsx")

#EPG data - time 
#br_time <- read.csv("./epg_data/Riviera comparison 8 hours/Riviera_Comparison_8_hours_timebins_0-8h.csv", header=TRUE)
br_time <- read.csv("./epg_data/Output/Rivera_timebins_0-8h.csv")

#EPG data - variables
#br_var <- read.csv("./epg_data/Riviera comparison 8 hours/Riviera_Comparison_8_hours_variables_0-8h.csv", header=TRUE)
br_table <- read.csv("./epg_data/Output/Rivera_epgtable_0-8h.csv")

#Variables
br_var <- read.csv("./epg_data/Output/Rivera_variables_0-8h.csv")


```

#cleaning df
```{r}

#Get info from file name
br_time2 <- br_time %>% 
            tidyr::separate(file, into = c("folder_path", "file_details"), sep = "Raw_data\\\\", remove = FALSE)

br_time3 <- br_time2 %>% 
            tidyr::separate(file_details, into = c("group_file", "rest"), sep = "/", remove = FALSE)

br_time4 <- br_time3 %>% tidyr::extract(
    rest,                       # Column to extract from
    into = c("day", "system", "channel"),  # New column names
    regex = "(\\d+)_syst(\\d+)_.*ch(\\d+)", # Regular expression to match the parts
    remove = FALSE                   # Keep the original column if needed
  )

br_time5 <- br_time4 %>% dplyr::select(-file, -folder_path,  -file_details, -group_file,  -rest) #ok now
br_time_ori <- br_time
br_time <- br_time5

#delete other df's
rm(br_time2, br_time3, br_time4, br_time5)

#make factors
br_time[ ,1:8] <- lapply(br_time[ ,1:8], as.factor)
str(br_time)

#make a new column named batch based on day
br_time2 <- br_time %>% 
           mutate(batch= case_when(
                    day %in% c("1", "2", "3", "4") ~ "batch_1",
                    day %in% c("5", "6", "7", "8") ~ "batch_2"))
time <- br_time2

#make new column a factor
time$batch <- as.factor(time$batch)

#check
head(time)
str(time)
summary(time)

x <- time %>% group_by(group, batch, wavename) %>% summarise(n=n())

#check levels
levels(time$wavename)
levels(time$group) #treatments
levels(time$id_rec) #rep number 
levels(time$wave)
levels(time$hour) #hours recording
levels(time$group)


#save as .rds data 
saveRDS(time, file="./Data/EPG_data/EPG_time.rds")
write.csv(time, file="./Data/EPG_data/csv/EPG_time.csv")



```

#Plots

## Time plots
```{r}
#Time spent on Phloem feeding & Pathway

#make a summary df only selecting for phloem feeding and pathway 
summary <- time %>%  
           filter(wavename %in% c("phloem feeding", "pathway", "xylem ingestion")) %>% 
           group_by(wavename, hour, group) %>%
           summarise(mean_time=mean(perc),
                     sd=sd(perc),
                     se=std.error(perc)) #from plotrix::std.error()

# Pathway ====

riv_pathway <- ggplot(subset(summary, wavename == "pathway"),
                   aes(x=hour, y=mean_time, color=group, group=group))+
    #geom_errorbar(aes(ymin=mean_time-se, ymax=mean_time+se), size=0.8, width=.2)+
         geom_line(stat="identity", linewidth=0.9)+
         geom_point(size=3)+
          geom_ribbon(aes(ymin = mean_time - se, ymax = mean_time + se, fill = group), # Shaded error area
              alpha = 0.2, color = NA) + 
         scale_color_manual(values= c("#BFBFBF", "#319B67"))+
         scale_fill_manual(values= c("#BFBFBF", "#319B67"))+
         ylim(0,80)+
         axis_looks+
        theme(legend.position="none")+
         labs(y="Time in pathway (%)", x="Time (h)")

riv_pathway 

#save plot
mapply(function(x)
  ggsave(
  plot = last_plot(),
  filename =x,
  path = "./Plots/05_EPG",
  scale = 1.8,
  width = 700,
  height = 600,
  units = "px",
  dpi = 300),
x=c("time_pathway.svg",
    "time_pathway.png"))

# Phloem ====

riv_phloem <- ggplot(subset(summary, wavename == "phloem feeding"),
                   aes(x=hour, y=mean_time, color=group, group=group))+
    #geom_errorbar(aes(ymin=mean_time-se, ymax=mean_time+se), size=0.8, width=.2)+
         geom_line(stat="identity", linewidth=0.9)+
         geom_point(size=3)+
          geom_ribbon(aes(ymin = mean_time - se, ymax = mean_time + se, fill = group), # Shaded error area
              alpha = 0.2, color = NA) + 
         scale_color_manual(values= c("#BFBFBF", "#319B67"))+
         scale_fill_manual(values= c("#BFBFBF", "#319B67"))+
         ylim(0,80)+
         axis_looks+
        theme(legend.position="none")+
         labs(y="Time phloem feeding (%)", x="Time (h)")

riv_phloem

#save plot
mapply(function(x)
  ggsave(
  plot = last_plot(),
  filename =x,
  path = "./Plots/05_EPG",
  scale = 1.8,
  width = 700,
  height = 600,
  units = "px",
  dpi = 300),
x=c("time_phloem.svg",
    "time_phloem.png"))


# Xylem ====
riv_xylem <- ggplot(subset(summary, plant_species == "Bole_Riv" & wavename == "xylem ingestion"),
                   aes(x=hour, y=mean_time, color=group, group=group))+
    #geom_errorbar(aes(ymin=mean_time-se, ymax=mean_time+se), size=0.8, width=.2)+
         geom_line(stat="identity", linewidth=0.9)+
         geom_point(size=3)+
          geom_ribbon(aes(ymin = mean_time - se, ymax = mean_time + se, fill = group), # Shaded error area
              alpha = 0.2, color = NA) + 
         scale_color_manual(values= c("#BFBFBF", "#4472C4"))+
         scale_fill_manual(values= c("#BFBFBF", "#4472C4"))+
         #ylim(0,60)+
         axis_looks+
        theme(legend.position="none")+
         labs(y="Time in xylem (%)", x="Time (h)")

riv_xylem

#save plot
mapply(function(x)
  ggsave(
  plot = last_plot(),
  filename =x,
  path = "./Plots/05_EPG",
  scale = 1.8,
  width = 700,
  height = 600,
  units = "px",
  dpi = 300),
x=c("time_xylem.svg",
    "time_xylem.png"))


```

## Aphid plot
```{r}

#call df for aphid population
aphid
str(aphid)
aphid$Plant_ID <- as.factor(aphid$Plant_ID)
aphid$Batch <- as.factor(aphid$Batch)
aphid$Time <- as.factor(aphid$Time)
aphid$Soil_stress <- as.factor(aphid$Soil_stress)
aphid$Soil_species <- as.factor(aphid$Soil_species)
aphid$Soil_treatment <- as.factor(aphid$Soil_treatment)

#plot aphid population boxplots

# Rivera ====
riv_aphids <- ggplot(subset(aphid, Soil_species == "Riv"),
                   aes(x=Soil_stress, y=aphids_14dpi, fill=Soil_stress))+
         geom_boxplot(mapping = aes(fill= Soil_stress),
                      position = position_dodge(width = 0.8), lwd=0.8)+
         geom_jitter(subset(aphid, Soil_species == "Riv"),
              mapping = (aes(x=Soil_stress, y=aphids_14dpi, fill = Soil_stress)),
              shape = 21, size = 1.9, alpha=0.9, color = "black",
              position = position_jitterdodge(jitter.height = .1, jitter.width = .2))+
         scale_fill_manual(values= c("#BFBFBF", "#4472C4"))+
         scale_x_discrete(labels=c("Ctr" = "Control-inocula", "SA" = "SA-inocula"))+
  ylim(0,60)+
         axis_looks+
  theme(legend.position="none")+
  labs(y="Number of M. persicae aphids 14dpi", x="")

riv_aphids

#save plot
mapply(function(x)
  ggsave(
  plot = last_plot(),
  filename =x,
  path = "./Plots/05_EPG",
  scale = 1.8,
  width = 700,
  height = 600,
  units = "px",
  dpi = 300),
x=c("aphid_number.svg",
    "aphid_number.png"))  
         

```

## Plant biomass plot
```{r}
plant
str(plant)
plant$Plant_ID <- as.factor(plant$Plant_ID)
plant$Batch <- as.factor(plant$Batch)
plant$Soil_stress <- as.factor(plant$Soil_stress)
plant$Soil_species <- as.factor(plant$Soil_species)
plant$Soil_treatment <- as.factor(plant$Soil_treatment)

#plot plant biomass boxplots

# Rivera ====
riv_shoot <- ggplot(subset(plant, Soil_species == "Bole_Riv"),
                   aes(x=Soil_stress, y=`Shoot weight (mg)`, fill=Soil_stress))+
         geom_boxplot(mapping = aes(fill= Soil_stress),
                      position = position_dodge(width = 0.8), lwd=0.8)+
         geom_jitter(subset(plant, Soil_species == "Bole_Riv"),
              mapping = (aes(x=Soil_stress, y=`Shoot weight (mg)`, fill = Soil_stress)),
              shape = 21, size = 1.9, alpha=0.9, color = "black",
              position = position_jitterdodge(jitter.height = .1, jitter.width = .2))+
         scale_fill_manual(values= c("#BFBFBF", "#4472C4"))+
         scale_x_discrete(labels=c("Ctr" = "Control-inocula", "SA" = "SA-inocula"))+
  ylim(0,650)+
         axis_looks+
  theme(legend.position="none")+
  labs(y="Shoot biomass (mg)", x="")

riv_shoot


#save plot
mapply(function(x)
  ggsave(
  plot = last_plot(),
  filename =x,
  path = "./Plots/05_EPG",
  scale = 1.8,
  width = 700,
  height = 600,
  units = "px",
  dpi = 300),
x=c("plant_biomass.svg",
    "plant_biomass.png"))      
         
```


##stats

```{r}

#Aphid number ====

#full df
aphid

#subset to Rivera
Riv <- subset(aphid, Soil_species == "Riv")
Riv %>% group_by(Soil_stress) %>% summarise(n=n()) #not balanced


glm <- glm(aphids_14dpi ~ Soil_stress,
              family=Gamma(link="log"),
              data=Riv)

#check
simulateResiduals(glm, plot=T) #ok

#get summary
number_glm <- car::Anova(glm, type="III", test.statistic="LR")


#Plant biomass ====
biomass <- subset(plant, Soil_species == "Bole_Riv")
biomass %>% group_by(Soil_stress) %>% summarise(n=n()) #not balanced


lm <- glm(`Plant weight` ~ Soil_stress, 
          family=gaussian,
          data=biomass)

#check
simulateResiduals(lm, plot=T) #ok

#get summary
biomass_table <- car::Anova(lm, type="III", test.statistic="LR")

#Phloem feeding time ====
time_df <- time %>%  
           filter(wavename %in% c("phloem feeding", "pathway"))

phloem <- time_df %>% filter(wavename == "phloem feeding")
phloem %>% group_by(group) %>% summarise(n=n()) #not balanced

hist(phloem$perc) #lot's of zero's

#---- Option with ZeroInflated ---- 
#when you have zero-inflated data (use pscl package)
library(pscl)
phloem$perc <- as.integer(phloem$perc) #pscl requires integer instead of numeric

m1 <- zeroinfl(perc ~ group * hour| id_rec, data = phloem)
m1_table <- car::Anova(m1, type="III", test.statistic="Chisq")

#check assumptions
# H0 hypothesis: Does model1 explain the same amount of variation as model0? If P value < 0.05, then answer is: No, model1 is different from model0 and thus a valid model!
mnull <- update(m1, . ~ 1) #null model
pchisq(2 * (logLik(m1) - logLik(mnull)), df = 3, lower.tail = FALSE) #Chi-Sqr test, p is < 0.05 so ok

#emmeans
emmeans<- emmeans(m1, pairwise~ group | hour,
                  adjust="FDR") 
plot(emmeans)
summary(emmeans$contrasts, infer = TRUE) #contrast by hour

#---- Optiom with glmer and poisson (best so far) ----
phloem.poiss <- glmer(perc ~ group*hour + (1|id_rec), 
               family = poisson, 
               data = phloem)

#simulateResiduals(phloem.poiss, plot=T) #not the best
car::Anova(poss, type="III", test.statistic="Chisq")

emmeans<- emmeans(phloem.poiss , pairwise~ group | hour,
                  adjust="FDR") 
plot(emmeans)

#---- Option with glmmTMB (not working) ----
library(glmmTMB)

m2 <- glmmTMB(
  perc ~ hour + group + (1 | id_rec), # Count process with random intercept
  zi = ~ 1,               # Zero-inflation process
  family = nbinom2,       # Use Poisson (or try `nbinom2` for overdispersion)
  data = phloem,
  verbose=TRUE)

summary(m2)

x <- glm(log(perc) ~ group, 
          family=Gamma,
          data=phloem)

#---- Option with lmer (not working, not linear) ----

library(lme4)
library(lmerTest)

lme_model.phloem <- lmerTest::lmer(perc ~ group * hour + (1|id_rec), data = phloem)
simulateResiduals(lme_model.phloem, plot=T) # not ok
summary(lme_model.phloem )
anova(lme_model.phloem )

lme_model.pathway <- lmerTest::lmer(perc ~ group * hour + (1|id_rec), data = pathway)
summary(lme_model.pathway)
anova(lme_model.pathway)


phloem$perc <- (phloem$perc/100) * 0.99+0.005

lme_model.phloem <- glmmTMB(perc ~ group * hour + (1|id_rec),
                          family= binomial(link = "logit"),
                          data = phloem)

simulateResiduals(lme_model.phloem, plot=T) # not ok
summary(lme_model.phloem )
anova(lme_model.phloem)

lme_model.pathway <- lmerTest::lmer(perc ~ group * hour + (1|id_rec), data = pathway)
summary(lme_model.pathway)
anova(lme_model.pathway)

#---- Option with wilxocon test, nothing is significant (??) ----

#using wilcoxon test
library(broom)
wilcox  <- phloem %>%
  group_by(hour) %>%
  do(tidy(wilcox.test(perc ~ group, data = .))) 

one <- phloem %>% subset(hour %in% "1")
wilcox.test(perc ~ group, exact = TRUE, data=one)

five <- phloem %>% subset(hour %in% "5")
wilcox.test(perc ~ group, exact = TRUE, data=five)



#Pathway time ====
pathway <- time_df %>% filter(wavename == "pathway")
pathway %>% group_by(group) %>% summarise(n=n()) #not balanced

hist(pathway$perc) #lot's of zero's also

#Option with ZeroInflated ====

pathway$perc <- as.integer(pathway$perc) #pscl requires integer instead of numeric

m1 <- zeroinfl(perc ~ group * hour| id_rec, data = phloem)
m1_table <- car::Anova(m1, type="III", test.statistic="Chisq")


#Option with Poisson ====
library(lme4)
path.poss <- glmer(perc ~ group*hour + (1|id_rec), 
               family = poisson, 
               data = pathway)

car::Anova(path.poss, type="III", test.statistic="Chisq")

emmeans<- emmeans(path.poss, pairwise ~ group | hour,
                  adjust="FDR") 
plot(emmeans)

#check
simulateResiduals(poss, plot=T) #ok

```



```{r Pathway; Stats}

#Pathway time ====
pathway <- time %>% filter(wavename == "pathway") %>% as.data.frame()
pathway %>% group_by(group) %>% summarise(n=n()) #not balanced
str(pathway)
pathway$perc <- as.numeric(pathway$perc)

#there is a conflict with lme4 package, you have to re-start R and just load lme4, Matrix and emmeans 
library(lme4)
library(Matrix)
library(emmeans)
library(pscl)
library(lmerTest)

#run lmer
lme_model.pathway <- lmerTest::lmer(perc ~ group * hour + (1|id_rec) + (1|batch), data = pathway)
lme_model.pathway.NoBatch <- lmerTest::lmer(perc ~ group * hour + (1|id_rec), data = pathway)
anova(lme_model.pathway, lme_model.pathway.NoBatch)  # Compare models, without batch is lower, so let's remove Batch

#summary
anova(lme_model.pathway.NoBatch)

#checking assumptions
residuals_lme <- residuals(lme_model.pathway.NoBatch)
plot(residuals_lme) #ok
qqnorm(residuals_lme) #ok
qqline(residuals_lme, col = "red") #ok
hist(residuals_lme, breaks = 20, main = "Histogram of Residuals") #there's a bi-modal shape

#let's see if transforming the data improves
#transform data
pathway$perc_log <- log(pathway$perc + 1)
pathway$perc_sqrt <- sqrt(pathway$perc)

#run lme again
# Refit the model with the transformed response
lme_model_log <- lmer(perc_log ~ group * hour + (1 | id_rec), data = pathway)
lme_model_sqrt <- lmer(perc_sqrt ~ group * hour + (1 | id_rec), data = pathway)

# Compare residuals
par(mfrow = c(1, 2))
hist(residuals(lme_model_log), main = "Log Transformed Residuals")
hist(residuals(lme_model_sqrt), main = "Sqrt Transformed Residuals") #yes!
dev.off()

#check again
residuals_lme <- residuals(lme_model_sqrt)
plot(residuals_lme)
qqnorm(residuals_lme)
qqline(residuals_lme, col = "red")
hist(residuals_lme, breaks = 20, main = "Histogram of Residuals") #bell-shaped now
boxplot(residuals_lme, main = "Boxplot of Residuals") #no outliers

#final summary
anova(lme_model_sqrt) #Satterthwaite's F test is prefered for fixed effects in LMMs

#post-hoc
emmeans(lme_model_sqrt, pairwise ~ group | hour, adjust="FDR") #ok

```




```{r}
pathway <- time %>%  
           filter(wavename %in% c("pathway"))

table(pathway$id_rec, pathway$hour)

library(ez)

results <- ezANOVA(
  data = pathway,
  dv = perc,
  wid = hour,
  within = .(group)
)

print(results)

phloem <- time %>%  
           filter(wavename %in% c("phloem feeding"))

table(phloem$id_rec, phloem$hour)

results <- ezANOVA(
  data = phloem,
  dv = perc,
  wid = hour,
  within = .(group)
)

print(results)

install.packages("lme4")

library(lme4)
library(lmerTest)

lme_model.phloem <- lmerTest::lmer(perc ~ group * hour + (1|id_rec), data = phloem)
simulateResiduals(lme_model.phloem, plot=T) # not ok
summary(lme_model.phloem )
anova(lme_model.phloem )

lme_model.pathway <- lmerTest::lmer(perc ~ group * hour + (1|id_rec), data = pathway)
summary(lme_model.pathway)
anova(lme_model.pathway)


phloem$perc <- (phloem$perc/100) * 0.99+0.005

lme_model.phloem <- glmmTMB(perc ~ group * hour + (1|id_rec),
                          family= binomial(link = "logit"),
                          data = phloem)

simulateResiduals(lme_model.phloem, plot=T) # not ok
summary(lme_model.phloem )
anova(lme_model.phloem)

lme_model.pathway <- lmerTest::lmer(perc ~ group * hour + (1|id_rec), data = pathway)
summary(lme_model.pathway)
anova(lme_model.pathway)
 


library(glmmTMB)

glmm_model <- glmmTMB(perc ~ group * hour + (1|id_rec),
                          family= nbinom2,
                          data = phloem)




```





```{r}
save.image("./R_Environments/EPG/EPG.Rdata ")
```
